### 2019-10-21 df6070f
Squashed commit of the following:
commit 679f9b545d384f4597240b52ad664c62d77b9362
Author: Silvana Ayala <silvana.ayala@nrel.gov>
Date:   Sat Oct 19 10:55:20 2019 -0600

    paper

commit 9c134c3bc90abc263416018516f2f145b47920c0
Author: Ayala <sayala@nrel.gov>
Date:   Fri Oct 18 12:26:36 2019 -0600

    last test paper

commit fc551f4a11e4dd75af22aef2455bf2cd0086cf4c
Author: Ayala <sayala@nrel.gov>
Date:   Fri Oct 18 12:25:29 2019 -0600

    paper

commit 9740162bba6b8d4d06cc1319c576208e7a5b1727
Author: Ayala <sayala@nrel.gov>
Date:   Fri Oct 18 12:24:06 2019 -0600

    paper test 3

commit edd
[…]
