+++ paper/paper.md
@@ -22,15 +22,20 @@ bibliography: paper.bib
 
 # Summary
 
-bifacial_radiance is a national laboratory and community-supported open source tool that provides a set of functions and classes for simulating the performance of bifacial photovoltaic systems. bifacial_radiance aims to provide reference implementations of models relevant to bifacial performance, including for example algorithms for sky irradiance simulation, DC power, and irradiance calculation at the modules. bifacial_radiance is an important component of a growing ecosystem of open source tools for solar energy  [@Holmgren2018].
+bifacial_radiance is a national-laboratory-developed, community-supported, open-source toolkit that provides a set of functions and classes for simulating the performance of bifacial photovoltaic (PV) systems. (Bifacial PV modules collect light on the front as well as the rear side.) bifacial_radiance automates calculations of PV system layout and performance to use along with the popular ray-tracing software tool RADIANCE [@Ward1994]. Specific algorithms include design and layout of PV modules, reflective ground surfaces, shading obstructions, and irradiance calculations throughout the system, among others. bifacial_radiance is an important component of a growing ecosystem of open-source tools for solar energy [@Holmgren2018].
 
-bifacial_radiance is developed on Github by contributors from academia, national laboratories, and private industry. Bifacial_Radiance is released copyrighted by the Alliance for Sustainable Energy with a BSD 3-clause license allowing permissive use with attribution. Bifacial_radiance is extensively tested for functional and algorithm consistency. Continuous integration services check each pull request on multiple platforms and Python versions. The bifacial_radiance API is thoroughly documented and detailed tutorials are provided for many features. The documentation includes help for installation and guidelines for contributions. The documentation is hosted at readthedocs.org as of this writing. Github’s issue trackers, a google group and StackOverflow tag provide venues for user discussions and help.
+bifacial_radiance is hosted on Github and PyPi, and it was developed by contributors from national laboratories, academia, and private industry. bifacial_radiance is copyrighted by the Alliance for Sustainable Energy with a BSD 3-clause license allowing permissive use with attribution. bifacial_radiance is extensively tested for functional and algorithm consistency. Continuous integration services check each pull request on multiple platforms and Python versions. The bifacial_radiance application programming interface (API) is thoroughly documented, and detailed tutorials are provided for many features. The documentation includes help for installation and guidelines for contributions. The documentation is hosted at readthedocs.org as of this writing. Github’s issue trackers, a Google group and StackOverflow tag provide venues for user discussions and help.
 
[…]
