### 2019-10-21 d32d519
describe how to build gencumsky in gencumsky/README.txt
