+++ bifacial_radiance/data/gencumsky/README.txt
@@ -5,3 +5,5 @@ This is directory contains the source for the gencumulativesky program.  To comp
 g++ -D_XOPEN_SOURCE *.cpp -lm -o gencumulativesky
 
 (The _XOPEN_SOURCE definition is required so that M_PI is defined when math.h is included - this will need changing depending on your compiler)
+
+bifacial_radiance update Oct 2019:  run python make_gencumskyexe.py from the command line to compile a platform-specific instance of gencumulativesky.exe
