### 2019-10-21 905ad77
Updated virtual sensor locations to correct x and z offsets and account for module thickness. Addresses issue #185
