+++ bifacial_radiance/main.py
@@ -1500,13 +1500,15 @@ class RadianceObj:
                 offsetfromaxis = zgap
                 tto = 0                
         #TODO: replace these with functions
+        
+        modulez = 0.020
 
         if text is None:
             
             if not cellLevelModuleParams:
                 try:
-                    text = '! genbox black {} {} {} '.format(name2,x, y)
-                    text +='0.02 | xform -t {} {} {} '.format(-x/2.0,
+                    text = '! genbox black {} {} {} {} '.format(name2, x, y, modulez)
+                    text +='| xform -t {} {} {} '.format(-x/2.0,
                                             (-y*Ny/2.0)-(ygap*(Ny-1)/2.0),
                                             offsetfromaxis)
                     text += '-a {} -t 0 {} 0'.format(Ny, y+ygap)
@@ -1526,7 +1528,7 @@ class RadianceObj:
                     cc = c['xcell']/2.0
                     print("Module was shifted by {} in X to avoid sensors on air".format(cc))
 
-                text = '! genbox black cellPVmodule {} {} 0.02 | '.format(c['xcell'], c['ycell'])
+                text = '! genbox black cellPVmodule {} {} {} | '.format(c['xcell'], c['ycell'], modulez)
                 text +='xform -t {} {} {} '.format(-x/2.0 + cc,
                                  (-y*Ny / 2.0)-(ygap*(Ny-1) / 2.0),
                                  offsetfromaxis)
@@ -1603,6 +1605,7 @@ class RadianceObj:
 
         moduleDict = {'x':x,
                       'y':y,
+                      'z':modulez,
                       'scenex': x+xgap,
                       'sceney': np.round(y*Ny + ygap*(Ny-1), 8),
                       'scenez': np.round(zgap + diam / 2.0, 8),
@@ -2018,7 +2021,8 @@ class RadianceObj:
                                   'clearance_height':trackerdict[theta]['clearance_height'],
                                   'azimuth':trackerdict[theta]['surf_azm'],
                                   'nMods': sceneDict['nMods'],
-                                  'nRows': sceneDict['nRows']}
+                                  'nRows': sceneDict['nRows'],
+                                  'modulez': scene.moduleDict['z']}
                 except KeyError:
                     #maybe gcr is passed, not pitch
                     sceneDict2 = {'tilt':trackerdict[theta]['surf_tilt'],
@@ -2026,7 +2030,8 @@ class RadianceObj:
                                   'clearance_height':trackerdict[theta]['clearance_height'],
                                   'azimuth':trackerdict[theta]['surf_azm'],
                                   'nMods': sceneDict['nMods'],
-                                  'nRows': sceneDict['nRows']}
+                                  'nRows': sceneDict['nRows'],
+                                  'modulez': scene.moduleDict['z']}
 
                 radfile = scene._makeSceneNxR(moduletype=moduletype,
                                              sceneDict=sceneDict2,
@@ -2062,7 +2067,8 @@ class RadianceObj:
                                       'clearance_height': trackerdict[time]['clearance_height'],
                                       'azimuth':trackerdict[time]['surf_azm'],
                                       'nMods': sceneDict['nMods'],
-                                      'nRows': sceneDict['nRows']}
+                                      'nRows': sceneDict['nRows'],
[…]
