### 2019-10-25 #1823 @melissawm
If no one is available, I can do it despite perhaps not being able to give
specific feedback on the ideas.

Em qui, 24 de out de 2019 16:07, Kristen Thyng <notifications@github.com>
escreveu:

> Hi @katyhuff <https://github.com/katyhuff> would you be up for editing
> this submission?
>
