### 2019-10-25 c2b6515
Fixed sensor offsets
