### 2019-10-29 #1823 @melissawm
@sgrieve would you be willing to review this for JOSS? Thanks!
