### 2019-10-29 #1823 @melissawm
@wholmgren would you be available to review this software paper for JOSS? Thank you!
