### 2019-11-01 #1823 @melissawm
@dalonsoa would you be willing/available to review this software paper for JOSS? Thank you.
