### 2019-11-05 #1823 @melissawm
Hi @dalonsoa , JOSS usually requires (at least) two reviewers, and currently we only have one associated with this review. If you are available, we would welcome your review! If that's the case, just let me know and I'll assign you as reviewer and open the REVIEW issue. Thank you!
