### 2019-11-05 #1865 @whedon
[REVIEW]: bifacial_radiance: a python package for modeling bifacial solar photovoltaic systems
**Submitting author:** @shirubana (<a href="http://orcid.org/0000-0003-0180-728X">Silvana Ayala Pelaez</a>)
**Repository:** <a href="https://github.com/NREL/bifacial_radiance/" target ="_blank">https://github.com/NREL/bifacial_radiance/</a>
**Version:** v0.3.3.1

@wholmgren ,@dalonsoa and @usethedata please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
