### 2019-11-06 NREL/bifacial_radiance#190 @dalonsoa
Improvement of installation instructions
Hi,

I'm performing the review of your submission to JOSS, and have a few concerns that I think need improvement, in this particular case, on the installation instructions.

In general, written installation instructions seem a bit limited and should be improved. Specific areas that should be reviewed are:

- I suspect the all instructions are somewhat outdated, not mentioning the simplest installation method using `pip install bifacial_radiance`, given that bifacial_radiance is available in PyPI and installs all dependencies (except RADIANCE itself, which is clearly stated in STEP 0 - Prerequisites).
- **Will the software work on all platforms?** STEP 2 seems to suggest that it will not, talking about moving to a specific location a .exe file, which is a Windows only file. What does it happen if you use another platform? Is that functionality still available? Will the software work at all? Is there any alternative to those Windows-only tools? It is absolutely fine if `bifacial_radiance` is a Windows-only software, of course, but if that is the case, it should be clearly stated.
- STEP 3 is obscure on what “scene” files are or why they need to be in a separate directory. Probably it will become clearer after improving the Package Overview part.
- The Youtube video instructions (good initiative!) are clearly only for Windows, but seem outdated too, for example installing PVlib manually using pip while it is now one of the dependencies listed in “requirements.txt”.

Best wishes,
Diego
