### 2019-11-06 #1865 @dalonsoa
Acceptance-blocker due to limited/incomplete installation instructions https://github.com/NREL/bifacial_radiance/issues/190
