### 2019-11-06 NREL/bifacial_radiance#191 @dalonsoa
Missing Package Overview page
The [Package Overview](https://bifacial-radiance.readthedocs.io/en/latest/package_overview.html) page in the documentation is missing (under construction). This is not a problem in itself, but it means that most of the high level description of what bifacial_radiance is and how it works is not there. Out of the information available in the [Intro Examples](https://bifacial-radiance.readthedocs.io/en/latest/introexamples.html), [ModelChain](https://bifacial-radiance.readthedocs.io/en/latest/modelchain.html) and the Jupyter Notebooks examples, it is very difficult to understand the workflow of the software, what the main elements that need to be created are or how the code is structured (keeping aside that ModelChain also seems to be largely a work in progress).

I would suggest to create some sort of flow diagram to illustrate this, indicating the inputs the outputs and what happens in between, all of it as general as possible, without entering into specific details of function calls and parameters.
