### 2019-11-06 #1865 @dalonsoa
Not exactly an acceptance-blocker, but certainly desirable to understand how the software works https://github.com/NREL/bifacial_radiance/issues/191
