### 2019-11-06 NREL/bifacial_radiance#192 @dalonsoa
Missing "intended audience" in the paper and the documentation
The problem you are trying to solve is more or less clearly stated, maybe slightly obscure for non-experts in PV, but it is reasonably clear. However, the intended audience is not.

Who will find `bifacial_radiance` useful? It is for students? For researchers? For PV system installers in a production environment?

A sentence on this will be useful to define more clearly who should be installing it.
