### 2019-11-06 #1865 @dalonsoa
Acceptance-blocker due to missing statement of need (or more specifically, intended audience) https://github.com/NREL/bifacial_radiance/issues/192
