### 2019-11-06 NREL/bifacial_radiance#193 @dalonsoa
Incomplete state of the field
In the paper, you mention that you have compared `bifacial_radiance` with other tools, indicating the papers where such comparison is published. That's great and not often done, but for the sake of completeness, it will be useful to elaborate a little bit on that in the paper itself to put this tool into context with other solutions without the need of reviewing other references. A paper should be roughly self-contained.

A few things that will be interested to know:
- Are there many tools comparable to `bifacial_radiance`?
- In which way is `bifacial_radiance` different (in other words, why you created it if there was already other tools doing the same)?
- Is there general agreement with those other tools?

The one who wants to know more details can now check the references you mention, but for those who don't want, at least they have a context in which to put `bifacial_radiance`.
