### 2019-11-06 #1865 @dalonsoa
Not an acceptance blocker, but it will be useful for the authors to elaborate https://github.com/NREL/bifacial_radiance/issues/193
