### 2019-11-07 #1865 @dalonsoa
@melissawm It seems there’s a figure missing in the Software Paper. It is present in the repository but not rendered when creating the PDF. I'm not sure of the reason for that.
