### 2019-11-07 #1865 @dalonsoa
@melissawm The license of the software is `BSD-3-Clause`, clearly stated in the corresponding file, but in the [DOI 10.11578/dc.20180530.16](https://doi.org/10.11578/dc.20180530.16) it appears as `BSD 2-clause "Simplified" License`. I am not sure if it will necessary to consolidate the license in both places or, if not possible, to add a note somewhere indicating when the license was changed from one to the other. Or it it matters at all!
