### 2019-11-07 #1865 @shirubana
Well the DOI paper was published in 2017 when we started and it seems we changed the license to BSD-3 1 year ago, as per the github history of the file... @cdeline can you confirm if BSD-3 is what we will go on forever now and/or if we need to consolidate?

I think a simple solution would be to write a note of the switch on the documentation (maybe on the updates, package overview and/or contribution guidelines).
