### 2019-11-07 #1865 @cdeline
Yes, we’re using a BSD-3 license for the software for now and forever more.

Chris Deline
National Renewable Energy Laboratory
(303) 384-6359
