### 2019-11-07 #1865 @wholmgren
@melissawm the COI policy states:

> In addition, your recent (past year) association with the same organization of a submitter is a COI, for example, being employed at the same institution.

@shirubana was a graduate student at my organization, University of Arizona, from 2012- Dec 2018. So there is 1-2 month overlap over the past year. But, the COI policy also states as an example of when a waiver can be granted:

> Or if you and a submitter are both employed by the same very large organization but in different units without any knowledge of each other.

We knew of each other but very rarely interacted. Different departments, siloed research programs.

I wouldn't have any problem providing an impartial review of this paper.
