### 2019-11-07 #1865 @wholmgren
Also on the subject of the license, it includes this NREL specific language:

>NOTICE: This software was developed at least in part by Alliance for Sustainable
Energy, LLC (“Alliance”) under Contract No. DE-AC36-08GO28308 with the U.S.
Department of Energy and the U.S. Government retains for itself and others acting
on its behalf a nonexclusive, paid-up, irrevocable worldwide license in the software
to reproduce, prepare derivative works, distribute copies to the public, perform
publicly and display publicly, and to permit others to do so.

Does this violate the requirement "Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?" My own opinion is "no" but seems like a question for the editors.
