### 2019-11-07 #1865 @danielskatz
@wholmgren - the conflict you describe is weak enough that it can be waived, but thank you for making this possible issue transparent.
