### 2019-11-12 6e7718d
Fixed sensor xstart/zstart. The initial position should be offset by 1/2 of the increment value, not a full increment. Addresses issue #185
