### 2019-11-12 #1865 @melissawm
Thank you all for your patience while I was checking this. As far as the change in licensing, a note in the documentation would be ok.

However, as far as the NOTICE mentioned above by @wholmgren , the best solution would be to remove it from the licence file (mantaining it a pure BSD 3 license) and adding this note somewhere in the project (for example, in the README).

@shirubana do you think this is feasible?

Please let me know if you have further questions or comments.
