### 2019-11-12 #1865 @shirubana
@cdeline would this option of using the README instead of the License file to have the sponsoring/DOE acknowledgment work to satisfy NREL's legal stuff?
