### 2019-11-13 #1865 @cdeline
Hi everybody!  I'm investigating these two points with the legal team here at NREL, but for everyone's information, a LICENSE file wasn't added until v0.2.3 on 11/14/2018, and it was initiated as a BSD-3 license, and has remained the same since then.  So, I believe the DOI entry is incorrect and should probably be updated - there are a couple of other inaccuracies there including that we were supported by the CEC.
