### 2019-11-18 NREL/bifacial_radiance#190 @shirubana
Hello @dalonsoa ,

I have updates to the sphinx documentation and the JOSS journal as requested on this and other threads. Am I supposed to wait until all the review process is finished to push this commits/changes or do it now/alongside the comments to start answering those?

Thank you

Silvana
