### 2019-11-18 #1865 @cdeline
Hi all,  i've been given approval by NREL legal to replace LICENSE.txt with a stock BSD-3 from the OSI webpage.  I'm merging that in Master now.
