### 2019-11-19 NREL/bifacial_radiance#190 @dalonsoa
Hi Silvana,

I think the best option will be to open a pull request with the changes, indicating which issues are solved by that PR and how. This way, it is much easier to see the changes compared to the original version. Tag me and @wholmgren as reviewers and when we accept it, that means we are happy with the corrections. Would that work for you?

Best wishes,
Diego
