### 2019-11-19 #1865 @melissawm
Thank you, @cdeline and @shirubana!

@wholmgren and @dalonsoa this means the review can resume as normal. If you have any further questions, don't hesitate to ask!
