### 2019-11-19 NREL/bifacial_radiance#190 @shirubana
Sounds good! I'll proceed that way. Thank you.
