### 2019-11-19 3fe059c
Addressing #190
Improvement of installation instructions #190:
-Mention of pip install option
-Added section dedicated to compatibility -- linux based systems OK
-Step 2 on install mentiosn how to install for Linux/Mac OSX steps.
-Step 3 phrasing on 'Scene' was modified/expanded for clarity.
-Youtube's description now includes updates to mention how to pip install bifacial_radiance, pip install pvlib, and comments for Linux/Mac OSX users to read the install guide on readthedocs.
