+++ docs/sphinx/source/installation.rst
@@ -3,6 +3,12 @@
 Installation
 ============
 
+Compatibility
+~~~~~~~~~~~~~
+
+bifacial_radiance is coded and tested in Windows, but can also work on Linux and Mac OSX, particularly after improvements in the latest release :ref:`whatsnew_0302` which solved some of the binary issues for gencumsky. However, the functionalities in Linux are still being improved, for example the GUI requires special QT installation described in (:issue:`130`:).
+
+
 Video Instructions
 ~~~~~~~~~~~~~~~~~~
 
@@ -21,7 +27,13 @@ If you are on a PC you should also copy the `Jaloxa radwinexe-5.0.a.8-win64.zip 
   
 STEP 1
 ~~~~~~
-Install and Import bifacial_radiance
+
+The simplest option is to open a command prompt and run::
+
+        pip install bifacial_radiance
+        
+       
+An alternative which is shown in the Video Instructions, if you want to install bifacial_radiance in a local folder of your choosing and/or be able to modify the internal code to suit your needs, you can do the following:
 
 * clone the bifacial_radiance repo to your local directory or download and unzip the .zip file
 * navigate to the \bifacial_radiance directory using anaconda command line
@@ -38,11 +50,17 @@ For best compatibility, deploy in an `Anaconda 2.7` environment, or run::
 
 STEP 2
 ~~~~~~
-Move gencumulativesky.exe
+Windows:
 
 * Copy gencumulativesky.exe from the repo's ``/bifacial_radiance/data/`` directory and copy into your Radiance install directory.
   This is typically found in ``/program files/radiance/bin/``.  
  
+Linux/Mac OSX:
+
+* A make_gencumskyexe.py script that builds gencumsky executable on pultiple platforms (Linux, Mac OSX, and Windows too) is included in bifacial_radiance release. More details on the use of this script on readme.txt or on thread (:issue:`1821`).
+* For Linux/Mac OSX, you will need to install QT for the GUI to work properly. Installation and details described in (:issue:`130`:).
+
+
 .. note::
         GenCumulativeSky is detailed in the publication "Robinson, D., Stone, A., Irradiation modeling made simple: the cumulative sky approach and its applications, Proc. PLEA 2004, Eindhoven 2004."   
 
@@ -51,10 +69,10 @@ The source is `available from the authors here <https://documents.epfl.ch/groups
 
 STEP 3
 ~~~~~~
-Create a local Radiance directory for storing the scene files created
+Create a local directory for storing your simulations and runs results. 
+If run in the default directory, simulation results will be saved in the TEMP folder, but will also be overwritten with every run. We recommend to keep the simulation files (scene geometry, skies, results, etc) separate from the bifacial_radiance directory by creating a local directory somewhere to be used for storing those files.
 
-Keep scene geometry files separate from the bifacial_radiance directory.  Create a local directory somewhere to be used for storing scene files.
- 
+
 STEP 4
 ~~~~~~
 Reboot the computer
