### 2019-11-19 8b7c683
Addresses #192, #193
Addressing #192 Intended Audience:
-Modified index page of documentation to include intended audience
-Modified paper.md to include intended audience.

Addressing #193 State of hte Field:
-Added state of hte field description, with tools comparable to bifacial_radiance and agreement in paper.md
-Still need to address pvfactors and pvlib mention and references.
