### 2019-11-19 c0334b1
Addressing #194
Addresing #194 Jupyter dependency:
-Mentioned on installation, that it is suggested to install the latest Anaconda to also get Jupyter for use with the tutorials.
-Added the code snippet to setup.py
