### 2019-11-19 f21b502
update Linux installation and make_gencumsky note.
