### 2019-11-19 cd709aa
move make_gencumsky directions into optional note.
