### 2019-11-27 76c2767
pytest  getEPW
added test for lines 424-495 (71 continuous lines) for getEPW. 2% improvement in coverage of Main (at 62% now.
