### 2019-11-27 bedea2a
remove HPCExample from main.py, add HPCExample Tutorial Notebook
remove HPCExample from main.py, add HPCExample Tutorial Notebook.
It was not working as an internal function nayways it was mostly for demonstration. Now as a journal it almost does stuff... still need to debug why it's not going into runJob().
Improves coverage by 2% (now at 64!)
