+++ bifacial_radiance/main.py
@@ -3665,114 +3665,7 @@ def runJob(daydate):
                                      modWanted=modWanted,
                                      rowWanted=rowWanted,
                                      sensorsy=sensorsy)
-
-def hpcExample():   
-    """ 
-    Example of HPC Job call
     
-    This allocates the day_dates generated to the different codes in as many nodes are available. 
-    Works inside and outside of slurm for testing (but set FullYear to False so it only does two days)
-    Full year takes 1 min in 11 Nodes.
-           
-    Variables stored in input_bf.py. First configure this on top:
-    
-    .. code-block :: python
-    
-        if __name__ == "__main__": #in case this is run as a script not a module.
-            from readepw import readepw  
-            from load import loadTrackerDict
-            from input_bf import *
-
-        else: # module imported or loaded normally
-            from bifacial_radiance.readepw import readepw # epw file reader from pvlib development forums  #module load format
-            from bifacial_radiance.load import loadTrackerDict
-            from bifacial_radiance.input_bf import *
-                
-    Procedure for a Full Year Run (~1 min in 11 nodes of 36 cores each > 365 days):
-
-    .. code-block :: python
-
-        Connect to Eagle
-        - $ cd bifacial_radiance/bifacial_radiance
-        - $ srun -A pvsoiling -t 5 -N 11 --pty bash                  
-        - $ module load conda
-        - $ . activate py3
-        - $ srun bifacial_radiance2.py
-    
-    Procedure for testing before joining SLURM:
-    
-    .. code-block :: python
-
-        change fullYear to False.
-        - $ cd bifacial_radiance/bifacial_radiance
-        - $ module load conda
-        - $ . activate py3
-        - $ nano bifacial_radiance.py    
-        - $ python bifacial_radiance2.py       
-    
-    .. warning::
-        Do not load conda twice nor activate .py3 twice.
-        (following above) Either activate conda or .py3 in the login node
-        or on the slurm
-   
-    """
-    
-    import multiprocessing as mp
-
-    daylist = []
-
-    fullYear = True # running faster testing on HPC ~ only 2 days.
-
-    if fullYear:
-        start = datetime.datetime.strptime("01-01-2014", "%d-%m-%Y")
-        end = datetime.datetime.strptime("31-12-2014", "%d-%m-%Y") # 2014 not a leap year.
-        daylist.append('12_31')     # loop doesn't add last day. Adding it at the beginning because why not.
-        daylimit = 365
-    else:
-        start = datetime.datetime.strptime("14-02-2014", "%d-%m-%Y")
-        end = datetime.datetime.strptime("26-02-2014", "%d-%m-%Y") # 2014 not a leap year.
-        daylimit = 1
-    date_generated = [start + datetime.timedelta(days=x) for x in range(0, (end-start).days)]
-    for date in date_generated:
-        daylist.append(date.strftime("%m_%d"))
-
-    #  print("This is daydate %s" % (daydate))
-    demo = RadianceObj(simulationname,path=testfolder)
-    demo.setGround(albedo)
-#   HPC IMPORTANT NOTE:
-#   Multiple Nodes get confused when trying to write the JSON at the same time,
-#    so make sure moduletype is created before running slurm job for it to work.
-#   2 DO: Fix at some point of course.
-#    moduleDict=demo.makeModule(name=moduletype,x=x,y=y,bifi=bifi,
-#                           torquetube=torqueTube, diameter = diameter, tubetype = tubetype,
-#                           material = torqueTubeMaterial, zgap = zgap, numpanels = numpanels, ygap = ygap,
-#                           rewriteModulefile = True, xgap=xgap,
[…]
