### 2019-11-27 856049c
quickExample fix #201 and pytest
adding testfolder optional input to quickExample() so it can be used in the pytest, and also returning analysis so it can be used in the pytest. Also calls now the interative directory from main.py instead of re-defining it inside.
Solves #201

Added pytest
