### 2019-11-27 f2297f2
pytest TorqueTubes on Modules
Improves coverage by 2%, by making different modules with torquetubes of diff. shapes. Ta-da :D
