### 2019-11-27 1252917
added test_1axis_gencumSky()
Improves coverage by another 1% :D now at 68%
