### 2019-11-27 08a579e
makeSCene1axis test for sceneDict with Pitch and with GCR (#204 test)
pytest improvement, doesn't solve but helps identify if #204 is addressed.
