### 2019-11-27 9d1c6fa
pytest Torquetubes smallfix
assertion was too long for Python 2.7, so reduced it :)
