### 2019-11-27 a039529
pytest gendaylit2manual, small fix test_1axis_genCumSky()
added gendaylit2manual test

small fix test_1axis_genCumSky(): I removed the ygap of the makeModule so the assertion didn't work. Added it back
