### 2019-11-27 2f05732
add test for makeImage and makeFalseColor, small fix gendaylit2manual test
