### 2019-11-27 98932d3
pytest sceneDict HEIGHT options on makeScene1axis
added some validations for various height input options. it'd be awesome if we deprecated all of this options and just made it obligatory to use hub_height and clearance_height,... it would trim many dumb lines of code and warnings.
