### 2019-11-27 e0af625
added small pytests
added pytests for exportTrackerDict(),
and makeModule() (empty input),
and added hpc=True on makeOct and makeImage to go over those lines as well.
