### 2019-11-27 c5cafa0
pytest getSingleTimestampTrackerAngle
