### 2019-11-27 c2386c8
Removed deprecated nMods and nRows inputs and warning to makeScene1axis
Deprecation warning on makeScene1axis has been there for a long time. Time to remove it. Addresses #208 for next release.
