### 2019-12-05 #1865 @dalonsoa
I'm waiting for some changes to be made in this PR https://github.com/NREL/bifacial_radiance/pull/199 before I can tick all the remaining boxes in the review. It seems that the PR is not passing all checks for now, so no sure what is the status of the review.
