### 2019-12-05 27f4ef8
pytest smallfix
asserting file names varies between my computer and online due to skies/skyblablah or skies\\skyblablah so changed the assertion to just 'skies'
