### 2019-12-05 a452fd2
update to DOE citation #213
update to main DOE citation -- modified 11/18.
