+++ docs/sphinx/source/index.rst
@@ -28,7 +28,7 @@ Many of the contributors to bifacial_radiance work in institutions where
 citation metrics are used in performance or career evaluations. If you
 use bifacial_radiance in a published work, please cite:
 
-        Deline, Chris, Marion, William, and Ayala Pelaez, Silvana. Bifacial_Radiance. Computer Software. https://github.com/NREL/bifacial_radiance. California Energy Commission. 17 Dec. 2017. Web. doi:10.11578/dc.20180530.16.
+        Deline, Chris, and Ayala, Silvana. Bifacial_Radiance. Computer Software. https://github.com/NREL/bifacial_radiance. USDOE Office of Energy Efficiency and Renewable Energy (EERE), Solar Energy Technologies Office (EE-4S). 17 Dec. 2017. Web. doi:10.11578/dc.20180530.16. 
         https://www.osti.gov/doecode/biblio/6869
 
 Please also cite the DOI corresponding to the specific version of
