### 2019-12-05 9818770
update main citation #213
It's updated on JOSS but since not sure when we're merging, updating here.
