### 2019-12-05 ea8e752
pytest otherfix and notes
