### 2019-12-05 f5ebf9e
added citations paper.md #197 #193
Added citations pvlib, pvfactors, pvsyst and sam citations
