### 2019-12-05 af43110
platform and python version updates #198
updated paper, instead of adding the installer, for the moment. Created a issue-wishlist to add other platforms at a later time (#214)
