### 2019-12-05 3ee75cd
Update installation.rst
Addresses various comments on PR #199
