### 2019-12-05 d360efa
package overview #191
Added text and images to package overview. Also added link to Webinar tutorial as the package is very well explained there with animations
