+++ docs/sphinx/source/index.rst
@@ -13,7 +13,7 @@ The source code for bifacial_radiance is hosted on `github
 
 Please see the :ref:`installation` page for installation help.
 
-For examples on how to use bifacial_radiance, please see :ref:`package_overview` and our `Jupyter Notebook tutorials
+For examples on how to use bifacial_radiance, please see our `Webinar tutorial video <https://www.youtube.com/watch?v=1X9L-R-RVGA>`_. Also find more information on :ref:`package_overview` and our `Jupyter Notebook tutorials
 <http://github.com/Nrel/bifacial_radiance/docs/tutorials/>`_. The documentation assumes general familiarity with Python, NumPy, and Pandas. Google searches will yield many excellent tutorials for these packages.
 
 The bifacial_radiance GitHub wiki has a `Projects and publications that use bifacial_radiance <https://github.com/NREL/bifacial_radiance/wiki/>`_ page for inspiration and listing of your application.
