### 2019-12-05 #1865 @shirubana
@dalonsoa @danielskatz  I have addressed all of issues and the PR comments now as well.
