### 2019-12-06 #1865 @dalonsoa
I'll tick those things that have been solved... but the PR is still not passing 🤷‍♂ .
