### 2019-12-06 NREL/bifacial_radiance#217 @dalonsoa
Using cross-platform paths
With `bifacial_radiance` installed and fully up and running, I'm having some trouble with the Jupyter tutorials (very good, otherwise!) because of the paths use.

Sometimes, you set paths as `r'..\..\bifacial_radiance\TEMP'`, in others you use `r'../../bifacial_radiance/TEMP'` (the same but with forward slashes) and in other cases (for example, tutorial 5) you even use an absolute path `r'C:\Users\sayala\Documents\RadianceScenes\Demo3'`. This is very inconvenient for anyone trying to use the tutorials because they need to be replacing them everywhere.

Keeping aside the absolute path, which should be replaced by a relative one, I strongly suggest to use the `pathlib` library (part of the standard python library from python >=3.4) for cross platform paths compatibility. So in your code you will replace the above relative paths by (this is for Tutorial 1):

```python
from pathlib import Path

testfolder = Path().resolve().parent.parent / 'bifacial_radiance' / 'TEMP'
```

This will work no matter whether you use Windows, Linux or MacOS. But requires a modern python (>=3.4).
