### 2019-12-06 #1865 @dalonsoa
@shirubana Installing "Radiance" (the Step 0 dependency of bifacial_radiance) is a pain. That's absolutely not your fault and there's nothing you need to do there except, maybe, asking the "Radiance" team to improve the installation pipeline.

Other than an issue I've just opened with something more of a suggestion than a requirement (https://github.com/NREL/bifacial_radiance/issues/217), this is ready to go from my side - once all check pass in the open PR.
