### 2019-12-06 b090a15
small fix #217
Just journal 1 atm, testing it.
