### 2019-12-07 8c0f63e
installation PR fix
