### 2019-12-07 9047fdf
update installation.rst
