### 2019-12-10 d847937
Revert "Removed deprecated nMods and nRows inputs and warning to makeScene1axis"
This reverts commit c2386c86c61340ed0a9a636a81a1b6deced3033d.
