### 2019-12-10 9a2907c
Squashed commit of the following:
commit f46571f0ee158e77bb08b371639b73d924956cbb
Merge: 77f4d33 d847937
Author: cdeline <chris.deline@nrel.gov>
Date:   Tue Dec 10 14:25:27 2019 -0700

    Merge pull request #199 from NREL/master-JOSS

    Modifications to address documentation suggestions by JOSS reviewers

commit d84793764c9d9aa22b898d6b047621e240c70825
Author: cdeline <chris.deline@nrel.gov>
Date:   Tue Dec 10 14:18:06 2019 -0700

    Revert "Removed deprecated nMods and nRows inputs and warning to makeScene1axis"

    This reverts commit c2386c86c61340ed0a9a636a81a1b6deced3033d.

commit 1d
[…]
