### 2019-12-12 NREL/bifacial_radiance#193 @shirubana
Resolved in commits
