### 2019-12-12 NREL/bifacial_radiance#190 @shirubana
resolved in commits / PR
