### 2019-12-12 NREL/bifacial_radiance#192 @shirubana
Resolved in commit
