### 2019-12-12 NREL/bifacial_radiance#191 @shirubana
Resolved in commit
