### 2019-12-13 #1865 @wholmgren
@cdeline the license is still not quite standard BSD-3. The 3rd clause contains "the United States Government, the United States Department of Energy," and the disclaimer contains "IN NO EVENT SHALL THE COPYRIGHT HOLDER, CONTRIBUTORS, UNITED STATES GOVERNMENT OR UNITED STATES DEPARTMENT OF ENERGY, NOR ANY OF THEIR EMPLOYEES..." instead of "IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS...". It's not a blocker as far as I'm concerned, but thought I should mention it here in case anyone else objects.
