### 2019-12-13 #1865 @wholmgren
I'm not able to install radiance with a reasonable amount of effort on MacOS. I realize this is not  really the fault of the `bifacial_radiance` developers, and @dalonsoa was able to get it done, so I don't feel comfortable holding up the review on my account. I recommend 1. improving the MacOS install instructions or dropping the claim of support and 2. finding another reviewer. Seems that @mikofski is already spun up on radiance and could probably do the review pretty quickly. Sorry I can't be more help.
