### 2019-12-13 #1865 @shirubana
@CameronTStark you use bifacial_radiance/radiance on a mac, right? Any insights on installation issues/tricks that we should include in our instructions?
