### 2019-12-13 #1865 @CameronTStark
The installer for MacOS isn't descriptive about where it installs. I tried for a bit to find it but had an easier time just choosing a location by pressing the "Change Install Location..." button in the "Installation Type" stage of the install. Then I source it in my bash/zsh_profile like so:

`
export PATH=$HOME/bin/radiance/bin:$PATH
export RAYPATH=$HOME/bin/radiance/lib
export MANPATH=$HOME/bin/radiance/man
export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$HOME/bin/radiance/lib
export MDIR=$HOME/bin/radiance/lib
`
