### 2019-12-17 65df881
Small fix journal 6
modified sorting of tracker keys to demo.trackerdict since trackerdict is not being assigned on the first cell.

Also, modified enddate to 11/07, otherwise empty array.
