+++ bifacial_radiance/main.py
@@ -60,7 +60,7 @@ import os, datetime, sys
 from subprocess import Popen, PIPE  # replacement for os.system()
 import pandas as pd
 import numpy as np 
-
+from input import *
 
 # Mutual parameters across all processes
 #daydate=sys.argv[1]
@@ -674,14 +674,16 @@ class RadianceObj:
                                 'ghi':'GHI'
                                 }, inplace=True)
 
+        tempTMYtitle = 'epw_temp.csv'
         # Hpc only works when daydate is passed through. Daydate gives single-
         # day run option with zero GHI values removed.
         if daydate is not None: 
             dd = re.split('_|/',daydate)
             starttime = dd[0]+'_'+dd[1] + '_00'
             endtime = dd[0]+'_'+dd[1] + '_23'
+            tempTMYtitle = 'epw_temp_'+dd[0]+'_'+dd[1]+'.csv'
         
-        tmydata_trunc = self._saveTempTMY(tmydata,'epw_temp.csv', 
+        tmydata_trunc = self._saveTempTMY(tmydata,filename=tempTMYtitle, 
                                           starttime=starttime, endtime=endtime)
         if daydate is not None:  # also remove GHI = 0 for HPC daydate call.
             tmydata_trunc = tmydata_trunc[tmydata_trunc.GHI > 0]
@@ -1175,7 +1177,7 @@ class RadianceObj:
         count = 0  # counter to get number of skyfiles created, just for giggles
 
         trackerdict2={}
-        for i in range(startindex,endindex+1):
+        for i in range(startindex,endindex):
             time = metdata.datetime[i]
             filename = str(time)[5:-12].replace('-','_').replace(' ','_')
             self.name = filename
@@ -2076,7 +2078,8 @@ class RadianceObj:
 
 
     def analysis1axis(self, trackerdict=None, singleindex=None, accuracy='low',
-                      customname=None, modWanted=None, rowWanted=None, sensorsy=9):
+                      customname=None, modWanted=None, rowWanted=None, sensorsy=9,
+                      daydate=None):
         """
         Loop through trackerdict and runs linescans for each scene and scan in there.
 
@@ -2115,6 +2118,9 @@ class RadianceObj:
 
         if customname is None:
             customname = ''
+
+        if daydate is None:
+            daydate = ''
 
         if trackerdict == None:
             try:
@@ -2180,14 +2186,14 @@ class RadianceObj:
         # Save compiled results using _saveresults
         if singleindex is None:
         
-            print ("Saving a cumulative-results file in the main simulatoin folder." +
+            print ("Saving a cumulative-results file in the main simulation folder." +
                    "This adds up by sensor location the irradiance over all hours " +
                    "or configurations considered." +
                    "\nWarning: This file saving routine does not clean results, so "+
                    "if your setup has ygaps, or 2+modules or torque tubes, doing "+
                    "a deeper cleaning and working with the individual results "+
                    "files in the results folder is highly suggested.")
-            cumfilename = 'cumulative_results_%s.csv'%(customname)
+            cumfilename = 'cumulative_results_%s.csv'%(daydate)
             if self.cumulativesky is True: 
                 frontcum = pd.DataFrame()
                 rearcum = pd.DataFrame()
@@ -3628,30 +3634,38 @@ def runJob(daydate):
         #exit(1)
 
     print("entering runJob on node %s" % slurm_nnodes)
-    metdata = demo.readEPW(epwfile=epwfile, hpc=hpc, daydate=daydate)
+    
[…]
