### 2020-01-02 7a17c1b
AWS test files
test2 ist he most updated.
Delete this files when finishing with development....
