### 2020-01-06 4e1df7b
implement of #186 , fix #226
This seems to solve issue #226  for vertical modules.

Uses the code suggested on #185 and PR #186 , except for the xinc/2 and zinc/2. Does not seem to be necessary? At least for 4 sensors, it places it correctly at tilt = 0. Look at x,y and z rears (the front is hte one that has the offset of the module thickness). Also this particular fix would only work for E/W facing modules.
