+++ bifacial_radiance/main.py
@@ -1356,7 +1356,7 @@ class RadianceObj:
         return trackerdict
 
 
-    def makeModule(self, name=None, x=None, y=None, bifi=1, modulefile=None, text=None, customtext='',
+    def makeModule(self, name=None, x=None, y=None, z=None, bifi=1, modulefile=None, text=None, customtext='',
                    torquetube=False, diameter=0.1, tubetype='Round', material='Metal_Grey',
                    xgap=0.01, ygap=0.0, zgap=0.1, numpanels=1, rewriteModulefile=True,
                    axisofrotationTorqueTube=False, cellLevelModuleParams=None,  
@@ -1502,13 +1502,17 @@ class RadianceObj:
                 offsetfromaxis = zgap
                 tto = 0                
         #TODO: replace these with functions
-
+       
+        # Adding the option to replace the module thickess
+        if z is None:
+            z = 0.020
+            
         if text is None:
             
             if not cellLevelModuleParams:
                 try:
-                    text = '! genbox black {} {} {} '.format(name2,x, y)
-                    text +='0.02 | xform -t {} {} {} '.format(-x/2.0,
+                    text = '! genbox black {} {} {} {} '.format(name2,x, y, z)
+                    text +='| xform -t {} {} {} '.format(-x/2.0,
                                             (-y*Ny/2.0)-(ygap*(Ny-1)/2.0),
                                             offsetfromaxis)
                     text += '-a {} -t 0 {} 0'.format(Ny, y+ygap)
@@ -1528,7 +1532,7 @@ class RadianceObj:
                     cc = c['xcell']/2.0
                     print("Module was shifted by {} in X to avoid sensors on air".format(cc))
 
-                text = '! genbox black cellPVmodule {} {} 0.02 | '.format(c['xcell'], c['ycell'])
+                text = '! genbox black cellPVmodule {} {} {} | '.format(c['xcell'], c['ycell'], z)
                 text +='xform -t {} {} {} '.format(-x/2.0 + cc,
                                  (-y*Ny / 2.0)-(ygap*(Ny-1) / 2.0),
                                  offsetfromaxis)
@@ -1605,6 +1609,7 @@ class RadianceObj:
 
         moduleDict = {'x':x,
                       'y':y,
+                      'z':z,
                       'scenex': x+xgap,
                       'sceney': np.round(y*Ny + ygap*(Ny-1), 8),
                       'scenez': np.round(zgap + diam / 2.0, 8),
@@ -2006,7 +2011,8 @@ class RadianceObj:
                                   'clearance_height':trackerdict[theta]['clearance_height'],
                                   'azimuth':trackerdict[theta]['surf_azm'],
                                   'nMods': sceneDict['nMods'],
-                                  'nRows': sceneDict['nRows']}
+                                  'nRows': sceneDict['nRows'],
+                                  'modulez': scene.moduleDict['z']}
                 except KeyError:
                     #maybe gcr is passed, not pitch
                     sceneDict2 = {'tilt':trackerdict[theta]['surf_tilt'],
@@ -2014,7 +2020,8 @@ class RadianceObj:
                                   'clearance_height':trackerdict[theta]['clearance_height'],
                                   'azimuth':trackerdict[theta]['surf_azm'],
                                   'nMods': sceneDict['nMods'],
-                                  'nRows': sceneDict['nRows']}
+                                  'nRows': sceneDict['nRows'],
+                                  'modulez': scene.moduleDict['z']}
 
[…]
