### 2020-01-07 d455d8a
update installation.rst
Added Mac OSX Instructions w Quartz from  notes on MacBook pro usage from and Alternative Installation with Windows Subsystem for Linux. from @mikofski , and @CameronTStark github and email threads.
This adds information of users who have bifacial_radiance on other platforms to address https://github.com/openjournals/joss-reviews/issues/1865

Clarified that the gencumksy executable for linux is the one without the .exe that should be moved to the folder addressing issue on #223

Fix contradiction on #222.
