+++ docs/sphinx/source/installation.rst
@@ -15,36 +15,119 @@ Video Instructions
 `https://youtu.be/4A9GocfHKyM <https://youtu.be/4A9GocfHKyM>`_ This video shows how to install the bifacial_radiance software and all associated softwares needed for Windows. More info on the Wiki. Instructions for Windows and Linux-based OS are also shown below.
 
 
-PREREQUISITES (Step 0)
-~~~~~~~~~~~~~~~~~~~~~~~
-This software requires the previous installation of ``RADIANCE`` from https://github.com/NREL/Radiance/releases.
- 
-Make sure you add radiance to the system PATH so Python can interact with the radiance program
- 
+STEP 0: PREREQUISITES (Step 0)
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+RADIANCE
+--------
+
 Windows:
 
+This software requires the previous installation of ``RADIANCE`` from https://github.com/NREL/Radiance/releases.
+ 
+Make sure you add radiance to the system PATH so Python can interact with the radiance program.
+
 If you are on a Windows computer you should also copy the `Jaloxa radwinexe-5.0.a.8-win64.zip  <http://www.jaloxa.eu/resources/radiance/radwinexe.shtml>`_ executables into ``program files/radiance/bin``. This executables allow for some nifty visualization options of your generated scene inside of bifacial_radiance, like falsecolor images.
+
 
 Linux/Mac OSX:
 
 * For Linux/Mac OSX, you will need to install QT for the GUI to work properly. Installation and details described in (:issue:`131`:):
 
-1. Install ``qt5-default`` from Ubuntu using ``apt``,
-2. get the official Radiance 5.2 source tarball with auxiliary libraries ``rad5R2all.tar.gz`` from either `RADIANCE <https://www.radiance-online.org/download-install/radiance-source-code/latest-release>`_ online or `LBL <https://floyd.lbl.gov/radiance/framed.html>`_ - do _not_clone the GitHub repo as it doesn't have the auxiliary libraries which you may also need. Finally extract the tarball.
-3. you may also need to install ``csh`` and ``cmake`` 
-4. make directories where you want to install radiance, for example ``~/.local/opt/radiance``
-5. read the README and run ``./makeall install clean`` and choose where you want ``bin`` and ``lib``
+1. Install ``qt5-default`` from Ubuntu using ``apt``.
+2. Get the official Radiance 5.2 source tarball with auxiliary libraries ``rad5R2all.tar.gz`` from either `RADIANCE <https://www.radiance-online.org/download-install/radiance-source-code/latest-release>`_ online or `LBL <https://floyd.lbl.gov/radiance/framed.html>`_ - do _not_clone the GitHub repo as it doesn't have the auxiliary libraries which you may also need. Finally extract the tarball.
+3. You may also need to install ``csh`` and ``cmake`` 
+4. Make directories where you want to install radiance, for example ``~/.local/opt/radiance``. Some users have reported that the installer for MacOS isn't descriptive about where it installs, and they have an easier time just choosing a location by pressing the "Change Install Location..." button in the "Installation Type" stage of the install. Then they source it in the bash/zsh_profile like so::
+
+        export PATH=$HOME/bin/radiance/bin:$PATH
+        export RAYPATH=$HOME/bin/radiance/lib
+        export MANPATH=$HOME/bin/radiance/man
+        export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$HOME/bin/radiance/lib
+        export MDIR=$HOME/bin/radiance/lib
+   
+5. Read the README and run ``./makeall install clean`` and choose where you want ``bin`` and ``lib``
 
[…]
