+++ docs/tutorials/10 - Running a Comparison with PVSyst.py
@@ -1,0 +1,1338 @@
+#!/usr/bin/env python
+# coding: utf-8
+
+# In[9]:
+
+
+import os
+from pathlib import Path
+import bifacial_radiance
+
+
+# In[10]:
+
+
+#Weather File:
+epwfile1 = r'C:\Users\sayala\Documents\GitHub\bifacial_radiance\bifacial_radiance\TEMP\Site1.epw'
+epwfile2 = r'C:\Users\sayala\Documents\GitHub\bifacial_radiance\bifacial_radiance\TEMP\Site2.epw'
+epwfile3 = r'C:\Users\sayala\Documents\GitHub\bifacial_radiance\bifacial_radiance\TEMP\Site3.epw'
+
+
+# <a id='step0b'></a>
+
+# ## Case 1
+
+# In[3]:
+
+
+testfolder = Path().resolve().parent.parent / 'bifacial_radiance' / 'TEMP' / 'Case1'
+
+# Another option using relative address; for some operative systems you might need '/' instead of '\'
+# testfolder = os.path.abspath(r'..\..\bifacial_radiance\TEMP')  
+
+print ("Your simulation will be stored in %s" % testfolder)
+
+
+# In[4]:
+
+
+try:
+    os.mkdir(testfolder)
+except OSError:
+    print ("Folder maybe existed or else failed for other reason. Check.")
+    print(testfolder)
+else:
+    print ("Successfully created the directory %s " % testfolder)
+
+
+# In[7]:
+
+
+# Case 1, AUSTRALIA
+simulationName = 'Case_1'    # For adding a simulation name when defning RadianceObj. This is optional.
+moduletype = 'Custom-Module'    # We will define the parameters for this below in Step 4.
+epwfile = epwfile1
+lat = -34.91
+lon = 146.60
+
+# Scene variables
+nMods = 20
+nRows = 7
+hub_height = 1.3 # meters
+gcr = 0.305
+albedo = 0.2         
+bifi = 0.8
+
+# Traking parameters
+cumulativesky = False
+limit_angle = 52 # 72.1 # tracker rotation limit angle
+angledelta = 0.01 # we will be doing hourly simulation, we want the angle to be as close to real tracking as possible.
+backtrack = True 
+axisofrotationTorqueTube = False
+
+#Torquetube Params
+torqueTube = False
+diameter = 0.1
+tubetype = 'round'    # This will make an octagonal torque tube.
+torqueTubeMaterial = 'black'   # Torque tube of this material (0% reflectivity)
+zgap = 0
+
+#makeModule parameters
+x = 1
+y = 1.98
+xgap = 0.01
+ygap = 0
+numpanels = 1
+axis_azimuth = 187 
+
+sensorsy = 9
+#sensorsy =100 # HPC Value
+
+
+# ### Do one day of simulation to compare Hourly
+
+# In[8]:
+
+
+demo = bifacial_radiance.RadianceObj(simulationName, path = str(testfolder))  # Adding a simulation name. This is optional.
+demo.setGround(albedo) 
+metdata = demo.readWeatherFile(weatherFile = epwfile) 
+mymodule = demo.makeModule(name=moduletype, x=x, y=y,
+                torquetube=torqueTube, diameter=diameter, tubetype=tubetype, material=torqueTubeMaterial, 
+                xgap=xgap, ygap=ygap, zgap=zgap, numpanels=numpanels, 
+                axisofrotationTorqueTube=axisofrotationTorqueTube)
+trackerdict = demo.set1axis(metdata = metdata, limit_angle = limit_angle, backtrack = backtrack, 
+                            gcr = gcr, cumulativesky = cumulativesky)
+
[…]

[…]
