### 2020-01-07 395c76c
cumulative results save Try-except
I don't think this was the main issue with why the pytest was failing, but this helped in case the file can't be written for whatever reason (i.e. no trackerdict -- which was the issue.)
