### 2020-01-08 c97d798
installation updates
