### 2020-01-08 #1865 @shirubana
@melissawm touching base on this review, what are the next steps? We have improved the documentation for installation of Radiance as much as we can with what has worked for some of our users. Are we getting another reviewer? How do we move forward?

Thank you
