### 2020-01-20 ba74fc0
Glass addition
must assign z to be a smaller value otherwise it work weird. z= 0.0002 for example, for 200 nm
