### 2020-01-27 #1865 @melissawm
Hello all,

Due to the problems that @wholmgren faced during installation, I believe it would be best if another reviewer could step in and continue the review. I understand in this case there is sufficient documentation for the software in question, however the installation issues stem from a dependency, and that is not under the author's control.

I would like to thank @wholmgren for his work in this review and the authors for their patience. If you have any further questions, please let me know.
