### 2020-01-27 #1865 @melissawm
@mankoff, would you be willing to review this software paper for JOSS? Thanks in advance!
