### 2020-01-27 #1865 @melissawm
@shirubana I would like to point out that the license file is still not a standard BSD-3 license (as previously mentioned above). Can you please check that?
