### 2020-01-27 #1865 @cdeline
Hi.  I’m not sure what the hold-up on the BSD-3 license is exactly.  Github is not auto-detecting our license file as a BSD-3, but we are using the BSD-3 license template and this is a standard BSD-3 license.  Because our organization is ‘Alliance for Sustainable Energy, LLC’ aka NREL, and we’re a federally funded national lab, we are contractually obligated to include the US Government and Department of Energy as co-copyright holders, and therefore are included explicitly anywhere the copyright holder is referenced in the legal text.  Otherwise this is a bog-standard BSD-3 licence.

The reviewer (wholmgren) mentioned that this issue is not specifically a blocker.  I’m not a copyright lawyer, and this isn’t something that I can just unilaterally change- this was the guidance and language given by our legal team that US government and Department of Energy need to be included because they are co-copyright holders.

Thanks-

Chris Deline
National Renewable Energy Laboratory
(303) 384-6359
