### 2020-02-06 #1865 @melissawm
@cdeline thanks for the clarification. I believe in this case, this is ok. I'm sorry as I'm not familiar with the US government licensing terms.

I'll wait for a few more days and if we don't get an answer I'll ask another reviewer to complete the review.
