### 2020-02-14 #1865 @melissawm
@shirubana do you have any suggestions for reviewers from the potential reviewer's list for JOSS?
