### 2020-02-15 7aedd99
RGB albedos for Ground class
Can pass just 1 albedo value (i.e. albedo= 0.6), or values for RGB (i.e. albedo = [0.4, 0.3, 0.6] )
