### 2020-02-15 2d3e0ac
small fix testfolder path
with the new method to define the testfolder path for any OS, it needs to be str to work
