### 2020-02-15 881a04c
_savefile RGB option
save specific W/m2 for each channel
