### 2020-02-26 #1865 @shirubana
Sorry, now I dropped the ball. u_u

Based on prioritizing Python as language, having done more than 1 reviews in the past year and not being currently overwhelmed with research, and having some kind of atmospheric, energy, or optics experience: andreas-h, dhhagan, cmccomb, dhhagan, yxqd, sarats, usethedata, mikaem, leouieda
