### 2020-02-26 #1865 @melissawm
Hello @cmccomb! Would you be willing/available to review this software paper for JOSS?
