### 2020-03-11 fe5b0cc
indent and sort keys for module.json
