+++ docs/sphinx/source/whatsnew/v0.3.4.rst
@@ -1,0 +1,34 @@
+.. _whatsnew_0304:
+
+v0.3.4 (MM / DD / 2020)
+------------------------
+
+This is a major release that adds many power mismatch analysis features, GUI updates as well as documentation.
+
+**Deprecated support for Python 2.7. Stops support on this release(?)**
+
+API Changes
+~~~~~~~~~~~
+
+
+Enhancements
+~~~~~~~~~~~~
+* Enable ``module.json`` to be human readable and sorted. (:issue:`230`)
+
+Bug fixes
+~~~~~~~~~
+
+
+Testing
+~~~~~~~
+
+
+Documentation
+~~~~~~~~~~~~~~
+
+
+Contributors
+~~~~~~~~~~~~
+* Chris Deline (:ghuser:`cdeline`)
+* Silvana Ayala (:ghuser:`shirubana`)
+* Cameron Stark (:ghuser:`CameronTStark`)
