### 2020-03-14 #1865 @shirubana
Hello @arfon, we would still like to proceed if possible with this manuscript as best as we can. Our first reviewer @dalonsoa gave us much great feedback and we implemented all of it, but we've been stuck first with installation issues of Radiance raytrace software (not ours - we've documented as best we can from other users) and with getting another reviewer. We've put a lot of effort in this package, tutorials and associated material, as well as into the improvements for the JOSS journal and all of these efforts will continue -- this tool is very useful in our field. We've had many publications in the past based on this software and more forthcoming so, it would be great to have an official way such as JOSS for us and others to cite it, since we've found it's an issue that we have that many users don't know how to properly cite our package and refer to it. Please advise how to best move forward!

Thank you

Silvana Ayala
