### 2020-03-16 #1865 @melissawm
@usethedata would you be willing to review this software paper for JOSS? At this time, we understand the disruptions caused by the COVID-19 pandemic so if you are willing to do it (even if it takes more time than usual) we would greatly appreciate it if you could mention it here. Thanks!
