### 2020-04-02 #1865 @usethedata
Checking in.  It took me longer to get through some chaos in my world.  I'm starting in on my review.
