### 2020-04-02 #1865 @usethedata
This is not, IMO, a conflict of interest.  However, I am documenting it here for transparency.  I work at Oak Ridge National Laboratory, and I am therefore an employee of UT-Battelle, which is a 50/50 joint venture between the University of Tennessee and Battelle Memorial Institute.  The authors are at the National Renewable Energy Laboratory, which is another of the US Department of Energy labs, and are presumably employees of the Alliance for Sustainable Energy.  ASE is a joint venture between MRI Global and Battelle Memorial Institute.  Battelle Memorial Institute, therefore, has a financial interest in the legal entities which employ me and which employ the authors.  However, that is not a controlling interest in either entity. The authors and I are, for legal and practical purposes, employed by separate legal entities, which happen to share a common partial owner.
