### 2020-04-14 3e25b20
Revert "AWS test files"
This reverts commit 7a17c1b5799d5f3e17be3a4dfb4e9ad27cc7bf05.
