+++ bifacial_radiance/test2.py
@@ -1,3775 +1,0 @@
-#!/usr/bin/env python
-from __future__ import division  # avoid integer division issues.
-from __future__ import absolute_import # this module uses absolute imports
-"""
-@author: cdeline
-
-bifacial_radiance.py - module to develop radiance bifacial scenes, including gendaylit and gencumulativesky
-7/5/2016 - test script based on G173_journal_height
-5/1/2017 - standalone module
-
-Pre-requisites:
-    This software is written in Python 2.7 leveraging many Anaconda tools (e.g. pandas, numpy, etc)
-
-    *RADIANCE software should be installed from https://github.com/NREL/Radiance/releases
-
-    *If you want to use gencumulativesky, move 'gencumulativesky.exe' from
-    'bifacial_radiance\data' into your RADIANCE source directory.
-
-    *If using a Windows machine you should download the Jaloxa executables at
-    http://www.jaloxa.eu/resources/radiance/radwinexe.shtml#Download
-
-    * Installation of  bifacial_radiance from the repo:
-    1. Clone the repo
-    2. Navigate to the directory using the command prompt
-    3. run `pip install -e . `
-
-Overview:
-    Bifacial_radiance includes several helper functions to make it easier to evaluate
-    different PV system orientations for rear bifacial irradiance.
-    Note that this is simply an optical model - identifying available rear irradiance under different conditions.
-
-    For a detailed demonstration example, look at the .ipnyb notebook in \docs\
-
-    There are two solar resource modes in bifacial_radiance: `gendaylit` uses hour-by-hour solar
-    resource descriptions using the Perez diffuse tilted plane model.
-    `gencumulativesky` is an annual average solar resource that combines hourly
-    Perez skies into one single solar source, and computes an annual average.
-
-    bifacial_radiance includes five object-oriented classes:
-
-    RadianceObj:  top level class to work on radiance objects, keep track of filenames,
-    sky values, PV module type etc.
-
-    GroundObj:    details for the ground surface and reflectance
-
-    SceneObj:    scene information including array configuration (row spacing, clearance or hub height)
-
-    MetObj: meteorological data from EPW (energyplus) file.
-        Future work: include other file support including TMY files
-
-    AnalysisObj: Analysis class for plotting and reporting
-
-"""
-import logging
-logging.basicConfig()
-LOGGER = logging.getLogger(__name__)
-LOGGER.setLevel(logging.DEBUG)
-
-import os, datetime, sys
-from subprocess import Popen, PIPE  # replacement for os.system()
-import pandas as pd
-import numpy as np 
-from input import *
-
-# Mutual parameters across all processes
-#daydate=sys.argv[1]
-
-
-global DATA_PATH # path to data files including module.json.  Global context
-#DATA_PATH = os.path.abspath(pkg_resources.resource_filename('bifacial_radiance', 'data/') )
-DATA_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), 'data'))
-
-def _findme(lst, a): #find string match in a list. script from stackexchange
-    return [i for i, x in enumerate(lst) if x == a]
-
-
-def _normRGB(r, g, b): #normalize by each color for human vision sensitivity
-    return r*0.216+g*0.7152+b*0.0722
-
-def _popen(cmd, data_in, data_out=PIPE):
-    """
-    Helper function subprocess.popen replaces os.system
-    - gives better input/output process control
-    usage: pass <data_in> to process <cmd> and return results
-    based on rgbeimage.py (Thomas Bleicher 2010)
-    """
-    cmd = str(cmd) # gets rid of unicode oddities
-
[…]
