### 2020-04-14 0af07f1
Cherry-pick "HPC-fixes"
This reverts SOME of commit 48ef35b3e5ef4d8da2e9940ba2de1b6a7ec97d82.
