### 2020-04-14 a4d6c4f
Fix pytests by returning +1 index to gendaylit1axis
