### 2020-04-14 de9f2d9
minor documentation update in main.py
