+++ tests/ini_gencumsky.ini
@@ -1,0 +1,57 @@
+#Version 0.2.5b
+[simulationParamsDict]
+weatherFile: USA_CO_Boulder.724699_TMY2.epw
+getEPW: False
+simulationname: _test_fixedtilt_end_to_end
+moduletype: test
+rewriteModule: True
+cellLevelModule: False
+axisofrotationTorqueTube: False
+torqueTube: False
+hpc: False
+tracking: False
+cumulativeSky: True
+timestampRangeSimulation: False 
+daydateSimulation: True
+latitude: 37.5
+longitude: -77.6
+
+[timeControlParamsDict]
+HourStart: 13
+HourEnd: 13
+DayStart: 17
+DayEnd: 17
+MonthStart: 6
+MonthEnd: 6
+timeindexstart: 4020
+timeindexend: 4021
+
+[moduleParamsDict]
+numpanels: 1
+x: 1.59
+y: 0.95
+bifi: 0.90
+xgap: 0.0
+ygap: 0.0
+zgap: 0.0
+
+[sceneParamsDict]
+gcrorpitch: pitch
+gcr: 0.350
+pitch: 1.5
+albedo: 0.62
+nMods: 10
+nRows: 3
+azimuth: 180
+tilt: 10
+clearance_height: 0.2
+hub_height: 0
+
+
+[analysisParamsDict]
+sensorsy: 9
+modWanted: 5
+rowWanted: 2
+
+[torquetubeParamsDict]
+diameter: 0.0
