### 2020-04-14 3787cbd
doc whatsnew 0.3.3.1
added v0.3.3.1 and links to v0.3.3.1 and v0.3.3.2 rst documentation. v0.3.3.2 not updated yet.
