### 2020-04-15 f8e4ec7
Add exception handling around for loop in gendaylit1axis. Should fix #225
