### 2020-04-16 311d5e4
GUI: add timeIndexSimulation control. replace timestampRangeSimulation with timeIndexSimulation in all .ini files.  FIXES #236
