+++ bifacial_radiance/load.py
@@ -40,7 +40,7 @@ def load_inputvariablesfile(intputfile):
             tracking                  bool
             cumulativesky             bool
             daydateSimulation         bool
-            timestampRangeSimulation  bool
+            timeIndexSimulation  bool
             ========================  =======  =============================
     sceneParamsDict : Dictionary 
         gcrorpitch, gcr, pitch, albedo, nMods, nRows, 
@@ -74,7 +74,7 @@ def load_inputvariablesfile(intputfile):
     simulationControlDict = {'fixedortracked':ibf.fixedortracked,
                              'cumulativeSky': ibf.cumulativeSky,
                              'timestampSimulation': ibf.timestampSimulation,
-                             'timestampRangeSimulation': ibf.timestampRangeSimulation,
+                             'timeIndexSimulation': ibf.timeIndexSimulation,
                              'hpc': ibf.hpc,
                              'daydateSimulation': ibf.dayDateSimulation,
                              'singleKeySimulation': ibf.singleKeySimulation,
@@ -522,7 +522,7 @@ def readconfigurationinputfile(inifile=None):
     else:
         raise Exception("Missing sceneParams Dictionary! Breaking")
             
-    if simulationParamsDict['timestampRangeSimulation'] or simulationParamsDict['daydateSimulation']:
+    if simulationParamsDict['timeIndexSimulation'] or simulationParamsDict['daydateSimulation']:
         if config.has_section("timeControlParamsDict"):
             timeControlParamsDict2 = boolConvert(confdict['timeControlParamsDict'])
             timeControlParamsDict={} # saving a main dictionary with only relevant options.
@@ -656,10 +656,10 @@ def readconfigurationinputfile(inifile=None):
             timeControlParamsDict['HourStart']=int(timeControlParamsDict2['HourStart'])
             simulationParamsDict['daydateSimulation'] = True
             
-            if simulationParamsDict['timestampRangeSimulation']:
-                print("Load Warning: timestampRangeSimulation and daydatesimulation both set to True.",\
-                      "Doing daydateSimulation and setting timestampRangeSimulation to False")
-                simulationParamsDict['timestampRangeSimulation'] = False
+            if simulationParamsDict['timeIndexSimulation']:
+                print("Load Warning: timeIndexSimulation and daydatesimulation both set to True.",\
+                      "Doing daydateSimulation and setting timeIndexSimulation to False")
+                simulationParamsDict['timeIndexSimulation'] = False
                 
         except:
             try:
@@ -667,18 +667,18 @@ def readconfigurationinputfile(inifile=None):
                 timeControlParamsDict['MonthStart']=int(timeControlParamsDict2['MonthStart']) 
                 print("Load Warning: timecontrolParamsDict hourend / hourstart is wrong/nan",\
                       "but since valid start day and month values were passed, switching simulation to",\
-                      "daydatesimulation = True, timestampRangeSimulation = False")
+                      "daydatesimulation = True, timeIndexSimulation = False")
                 simulationParamsDict['daydateSimulation']=True
-                simulationParamsDict['timestampRangeSimulation']=False
+                simulationParamsDict['timeIndexSimulation']=False
             except:
                 print("Load Warning: no valid day, month and hour passed for simulation.",\
                       "setting cumulative to True, and daydatesimulation and ",\
-                      "timestampRangeSimulation to False")
+                      "timeIndexSimulation to False")
                 simulationParamsDict['cumulativeSky']=True
                 simulationParamsDict['daydateSimulation']=False
-                simulationParamsDict['timestampRangeSimulation']=False
+                simulationParamsDict['timeIndexSimulation']=False
     
[…]
