+++ bifacial_radiance/gui.py
@@ -10,6 +10,7 @@ except:
     import ttk
 
 import bifacial_radiance
+import warnings
 
 # aliases for Tkinter functions
 END = tk.END
@@ -87,7 +88,7 @@ class Window(tk.Tk):
                 
             testfolder, weatherfile, weatherinputMode, simulation,
             moduletype, rewriteModule, cellLevelModule, axisofrotationTorqueTube,
-            torqueTube, fixedortracking,  cumulativesky, timestampRangeSimulation,
+            torqueTube, fixedortracking,  cumulativesky, timeIndexSimulation,
             daydateSimulation, lat, lon, timestampstart, timestampend, entry_startdate_hour,
             entry_enddate_hour, entry_startdate_day, entry_enddate_day, entry_startdate_month,
             entry_enddate_month, numberofPanels, x, y, bifi, xgap, ygap, zgap, 
@@ -98,6 +99,7 @@ class Window(tk.Tk):
             
             
             '''
+            
             
             def _set_daily_endtimes():
                 '''
@@ -120,7 +122,7 @@ class Window(tk.Tk):
             
             testfolder, weatherfile, weatherinputMode, simulation,\
             moduletype, rewriteModule, cellLevelModule, axisofrotationTorqueTube,\
-            torqueTube, fixedortracking,  cumulativesky, timestampRangeSimulation,\
+            torqueTube, fixedortracking,  cumulativesky, timeIndexSimulation,\
             daydateSimulation, lat, lon, timestampstart, timestampend, enddate_day,\
             enddate_hour, startdate_day, startdate_hour, startdate_month,\
             enddate_month, numberofPanels, x, y, bifi, xgap, ygap, zgap, \
@@ -141,7 +143,7 @@ class Window(tk.Tk):
             
             # Initializing
             daydateSimulation = False
-            timestampRangeSimulation = False   
+            timeIndexSimulation = False   
             if rb_fixedortracking.get() == 0: 
                 fixedortracking=False # False, fixed. Fixed, Cumulative Sky Yearly
                 cumulativesky = True
@@ -149,7 +151,7 @@ class Window(tk.Tk):
                 fixedortracking=False # True, 'tracking' Fixed, Cumulative Sky with Start/End
                 cumulativesky = True
                 daydateSimulation = True #TODO: check this out. new 8/20/19
-                #timestampRangeSimulation = True
+                #timeIndexSimulation = True
             if rb_fixedortracking.get() == 2: 
                 fixedortracking=False # True, 'tracking'  Fixed, Hourly with Start/End times
                 cumulativesky = False
@@ -169,7 +171,7 @@ class Window(tk.Tk):
                 fixedortracking=True # True, 'tracking' Tracking, Hourly with Start/End times
                 cumulativesky = False
                 daydateSimulation = True #TODO: check this out. new 8/20/19
-                #timestampRangeSimulation = True
+                #timeIndexSimulation = True
             if rb_fixedortracking.get() == 7: 
                 fixedortracking=True # True, 'tracking' Tracking, Hourly for the Whole Year
                 cumulativesky = False
@@ -321,7 +323,7 @@ class Window(tk.Tk):
             simulationParamsDict['hpc'] =  False #Fix
             if fixedortracking is not None: simulationParamsDict['tracking'] =  fixedortracking
             if cumulativesky is not None: simulationParamsDict['cumulativeSky'] =  cumulativesky
-            if timestampRangeSimulation is not None: simulationParamsDict['timestampRangeSimulation'] =  timestampRangeSimulation
+            if timeIndexSimulation is not None: simulationParamsDict['timeIndexSimulation'] =  timeIndexSimulation
[…]
