+++ tests/ini_highAzimuth.ini
@@ -11,18 +11,12 @@ torqueTube: False
 hpc: False
 tracking: False
 cumulativeSky: False
-timestampRangeSimulation: True 
+timeIndexSimulation: True 
 daydateSimulation: False
 latitude: 37.5
 longitude: -77.6
 
 [timeControlParamsDict]
-HourStart: 13
-HourEnd: 14
-DayStart: 17
-DayEnd: 17
-MonthStart: 6
-MonthEnd: 6
 timeindexstart: 4020
 timeindexend: 4021
 
