+++ tests/ini_gencumsky.ini
@@ -11,7 +11,7 @@ torqueTube: False
 hpc: False
 tracking: False
 cumulativeSky: True
-timestampRangeSimulation: False 
+timeIndexSimulation: False 
 daydateSimulation: True
 latitude: 37.5
 longitude: -77.6
