### 2020-04-16 42621e3
Tutorial 12 added Goal AgriPV Image on description ad WHATSNEW v0.3.3.2
Tutorial 12 added Goal AgriPV Image on description ad WHATSNEW v0.3.3.2
