+++ docs/sphinx/source/whatsnew/v0.3.3.2.rst
@@ -1,155 +1,67 @@
 .. _whatsnew_030302:
 
-v0.3.3.2 (4 / 14 / 2020)
+v0.3.3.2 (04 / 16 / 2020)
 ------------------------
-
-This is a minor release with a handful of bug fixes and some additional functions
-
-**Deprecated support for Python 2.7. Will stop support on 0.3.4 release**
-
-
-
-@SILVANA - here are some additional notes to add:
-_tempTMY saves a file with the daydate now if available so it doesnt get overwritten/overreda.
-importnat fix to sky generation on gendaylit1axis, because the for loop was going to endindex+1 before.
-analysis1axis added input for daydate, so cumulative file adds for hte daydate. The other option is to turn off the cumulative file saving option for daydate? discuss
-modified sorting of tracker keys to demo.trackerdict since trackerdict is not being assigned on the first cell.
-* Removed deprecated nMods and nRows inputs and warning to makeScene1axis
-Deprecation warning on makeScene1axis has been there for a long time. Time to remove it. Addresses #208 for next release.
-JOSS updates:
-    Revert "Removed deprecated nMods and nRows inputs and warning to makeScene1axis"
-    remove extraneous tutorial1 .py file
-    update installation.rst
-    journal2 path fix #217
-    small fix #217
-    package overview #191
-    platform and python version updates #198
-    improvements to contributing #196
-    added citations paper.md #197 #193
-    Added citations pvlib, pvfactors, pvsyst and sam citations
-    citation on readme update #213
-    update to DOE citation #213
-    asserting file names varies between my computer and online due to skies/skyblablah or skies\\skyblablah so changed the assertion to just 'skies'
-    pytest getSingleTimestampTrackerAngle
-    pytest sceneDict HEIGHT options on makeScene1axis
-    added some validations for various height input options. it'd be awesome if we deprecated all of this options and just made it obligatory to use hub_height and clearance_height,... it would trim many dumb lines of code and warnings.
-    add test for makeImage and makeFalseColor, small fix gendaylit2manual test
-    pytest gendaylit2manual, small fix test_1axis_genCumSky()
-    added gendaylit2manual test
-    small fix test_1axis_genCumSky(): I removed the ygap of the makeModule so the assertion didn't work. Added it back
-    pytest Torquetubes smallfix
-    assertion was too long for Python 2.7, so reduced it :)
-    makeSCene1axis test for sceneDict with Pitch and with GCR (#204 test)
-    pytest improvement, doesn't solve but helps identify if #204 is addressed.
-    added test_1axis_gencumSky()
-    Improves coverage by another 1% :D now at 68%
-    pytest TorqueTubes on Modules
-    Improves coverage by 2%, by making different modules with torquetubes of diff. shapes. Ta-da :D
-    quickExample fix #201 and pytest
-    adding testfolder optional input to quickExample() so it can be used in the pytest, and also returning analysis so it can be used in the pytest. Also calls now the interative directory from main.py instead of re-defining it inside.
-    Solves #201
-    remove HPCExample from main.py, add HPCExample Tutorial Notebook
-    remove HPCExample from main.py, add HPCExample Tutorial Notebook.
-    pytest  getEPW
-    added test for lines 424-495 (71 continuous lines) for getEPW. 2% improvement in coverage of Main (at 62% now.
-    move make_gencumsky directions into optional note.
-    update Linux installation and make_gencumsky note.
-    Addresing #194 Jupyter dependency:
[…]
