### 2020-04-16 958c600
backwards compatibility load.py to accept timestampRangeSimulation key from old .ini file
