### 2020-04-16 d48f675
release notes 0.3.3.2. Pytest checks timestampRangeSimulation backward compatibility in .ini
