+++ docs/sphinx/source/whatsnew/v0.3.3.2.rst
@@ -1,0 +1,76 @@
+.. _whatsnew_030302:
+
+v0.3.3.2 (04 / 16 / 2020)
+------------------------
+Release of new version including fix to sensor orientation, increased py-testing, improved documentation, 
+
+
+Enhancements
+~~~~~~~~~~~~
+* Glass-front modules now supported.  New Glass material included in `ground.rad` and example for use in ``docs/tutorials/Glass Test.ipynb``
+* Source code for `gencumulativesky` is now included along with `makefile` instructions in ``/data/gencumsky/``
+* New Torquetube options added on py:class:`~bifacial_radiance.RadianceObj.makeModule`
+
+
+API Changes
+~~~~~~~~~~~
+* py:class:`~bifacial_radiance.RadianceObj.readTMY` now saves the temporal TMY files generated for daydate simulations with different titles to avoid overwriting in parallel simulations.
+* Input for daydate added to py:class:`~bifacial_radiance.RadianceObj.analysis1axis`, allowing to save cumulative files for daydate. 
+* Sorted trackerdict keys on simulations that use trackerdict routines
+* Removed deprecation warning and inputs ``nMods`` and ``nRows`` on py:class:`~bifacial_radiance.RadianceObj.makeScene1axis`
+* Modified ``quickExample`` to include testfolder as an input, and returned analysis results. Call to interactive directory selection called from `main.py` instead than being redefined inside the function. (:issue:`201`)
+* Removed ``HPCExample`` function from `main.py` (added as a Tutorial Notebook)
+* Fix GUI load issue when `None` values passed
+* Set GUI default ``modWanted`` and ``rowWanted`` to `Null`
+* .ini input files now use the key `timeIndexSimulation` instead of `timestampRangeSimulation`
+
+
+Bug fixes
+~~~~~~~~~
+* Fixed sensor’s orientation error in scanning function calculation, mainly affecting high-tilt angle simulations (:issue: `226`)
+* Fix GUI to allow reading null values for ``modWanted`` and ``rowWanted``
+* Exception handling for daily simulations in py:class:`~bifacial_radiance.RadianceObj.gendaylit1axis`, due to a for loop going out of range for endindex+1 (:issue:`225`)
+
+
+Testing
+~~~~~~~
+* Modified pytest assertions for skies to avoid OS issues (ski\etc or skie//etc)
+* Added pytesting to py:cass:`~bifacial_radiance.RadianceObj.getSingleTimestampTrackerAngle`
+* Added pytesting to `sceneDict` ``height`` in py:class:`~bifacial_radiance.RadianceObj.makeScene1axis`
+* Added pytesting for py:class:`~bifacial_radiance.AnalysisObj.makeImage` and py:class:`~bifacial_radiance.AnalysisObj.makeFalseColor`
+* Added pytest for py:class:`~bifacial_radiance.RadianceObj.gendaylit2manual` 
+* Added pytest for py:class:`~bifacial_radiance.RadianceObj.genCumSky` in py:class:`~bifacial_radiance.modelchain.runModelChain`
+* Added pytest for py:class:`~bifacial_radiance.RadianceObj.genCumSky1axis` on ``test_1axis_genCumSky()``
+* Added pytest for `torque tubes` py:class:`~bifacial_radiance.RadianceObj.makeModule`
+* Added pytest for `pitch` and `gcr` on py:class:`~bifacial_radiance.RadianceObj.makeScene1axis` (:issue:`204`)
+* Added pytest to ``quickExample`` (:issue:`201`)
+* Added pytest for py:class:`~bifacial_radiance.RadianceObj.getEPW`
+
+
+Documentation
+~~~~~~~~~~~~~~
[…]
