+++ tests/ini_gencumsky.ini
@@ -1,4 +1,6 @@
 #Version 0.2.5b
+#Specifically use timestampRangeSimulation key to test for backward compatibility
+# < 0.3.3.1
 [simulationParamsDict]
 weatherFile: USA_CO_Boulder.724699_TMY2.epw
 getEPW: False
@@ -11,7 +13,7 @@ torqueTube: False
 hpc: False
 tracking: False
 cumulativeSky: True
-timeIndexSimulation: False 
+timestampRangeSimulation: False 
 daydateSimulation: True
 latitude: 37.5
 longitude: -77.6
