### 2020-04-16 5a7ba6c
unicode character removal in whatsnew/v0.3.3.2.rst
