+++ docs/sphinx/source/whatsnew/v0.3.3.2.rst
@@ -1,7 +1,7 @@
-.. _whatsnew_030302:
+.. _whatsnew_0332:
 
 v0.3.3.2 (04 / 16 / 2020)
-------------------------
+--------------------------
 Release of new version including fix to sensor orientation, increased py-testing, improved documentation, 
 
 
@@ -22,12 +22,12 @@ API Changes
 * Removed ``HPCExample`` function from `main.py` (added as a Tutorial Notebook)
 * Fix GUI load issue when `None` values passed
 * Set GUI default ``modWanted`` and ``rowWanted`` to `Null`
-* .ini input files now use the key `timeIndexSimulation` instead of `timestampRangeSimulation`
+* py:class:`~bifacial_radiance.load.readconfigurationinputfile`.ini input files now use the key ``timeIndexSimulation`` instead of ``timestampRangeSimulation``
 
 
 Bug fixes
 ~~~~~~~~~
-* Fixed sensor’s orientation error in scanning function calculation, mainly affecting high-tilt angle simulations (:issue: `226`)
+* Fixed sensor's orientation error in scanning function calculation, mainly affecting high-tilt angle simulations (:issue:`226`)
 * Fix GUI to allow reading null values for ``modWanted`` and ``rowWanted``
 * Exception handling for daily simulations in py:class:`~bifacial_radiance.RadianceObj.gendaylit1axis`, due to a for loop going out of range for endindex+1 (:issue:`225`)
 
@@ -35,7 +35,7 @@ Bug fixes
 Testing
 ~~~~~~~
 * Modified pytest assertions for skies to avoid OS issues (ski\etc or skie//etc)
-* Added pytesting to py:cass:`~bifacial_radiance.RadianceObj.getSingleTimestampTrackerAngle`
+* Added pytesting to py:class:`~bifacial_radiance.RadianceObj.getSingleTimestampTrackerAngle`
 * Added pytesting to `sceneDict` ``height`` in py:class:`~bifacial_radiance.RadianceObj.makeScene1axis`
 * Added pytesting for py:class:`~bifacial_radiance.AnalysisObj.makeImage` and py:class:`~bifacial_radiance.AnalysisObj.makeFalseColor`
 * Added pytest for py:class:`~bifacial_radiance.RadianceObj.gendaylit2manual` 
@@ -50,7 +50,7 @@ Testing
 Documentation
 ~~~~~~~~~~~~~~
 * Updated bifacial_radiance D.O.E. reference citation on :ref:`_index` (:issue:`213`)
-* New Tutorial on AgriPV example included in ‘‘/docs/tutorials/12 AgriPV Systems.ipynb’’
+* New Tutorial on AgriPV example included in ''/docs/tutorials/12 AgriPV Systems.ipynb''
 * New Tutorial on running HPC Example
 * Added Webinar Slides and Tutorial Journal
 * Added `pip install` options to :ref:`_installation` 
@@ -61,7 +61,7 @@ Documentation
 * Platform and python-version updates throughout documentation (:issue:`198`)
 * Improvements to :ref:`_contributing` guidelines (:issue:`191`)(:issue:`196`)
 * Working Folder path fixed on Tutorial Journals for compatibility with other OS Systems (:issue:`217`)
-* Updated Youtube’s installation instructions to mention how to pip install bifacial_radiance, pip install pvlib, and comments for Linux/Mac OSX users to read the install guide on readthedocs.
+* Updated Youtube's installation instructions to mention how to pip install bifacial_radiance, pip install pvlib, and comments for Linux/Mac OSX users to read the install guide on readthedocs.
 * Added suggestion to install Anaconda to run Jupyter Tutorial notebooks in :ref:`_installation` (:issue:`194`)
 
 
