### 2020-04-16 31ae7d1
0332 release (#235)
* add patched gencumsky source, license, makefile

* remove logging of linker_exe and compiler_cxx, don't exist on windows

* Squashed commit of the following:

commit 679f9b545d384f4597240b52ad664c62d77b9362
Author: Silvana Ayala <silvana.ayala@nrel.gov>
Date:   Sat Oct 19 10:55:20 2019 -0600

    paper

commit 9c134c3bc90abc263416018516f2f145b47920c0
Author: Ayala <sayala@nrel.gov>
Date:   Fri Oct 18 12:26:36 2019 -0600

    last test paper

commit fc551f4a11e4dd75af22aef2455bf2cd0086cf4c
Author: Ayala <sayala@nrel.gov>
Date:   Fri Oct 18 12:25:29 2019 -0600

    paper

c
[…]
