+++ setup.py
@@ -110,7 +110,7 @@ setup(
     # $ pip install -e .[dev,test]
     
     extras_require={
-        'examples': ['jupyter'],
+        #'examples': ['jupyter'],
         #'dev': ['check-manifest'],
         #'test': ['coverage'],
     },
