### 2020-04-16 5dacc95
whatsnew & doc fixes
