+++ docs/sphinx/source/whatsnew/v0.3.3.2.rst
@@ -7,7 +7,7 @@ Release of new version including fix to sensor orientation, increased py-testing
 
 Enhancements
 ~~~~~~~~~~~~
-* Glass-front modules now supported.  New Glass material included in `ground.rad` and example for use in ``docs/tutorials/Glass Test.ipynb``
+* Glass in modules now supported. New Glass material included in `ground.rad`. Example of use in `Glass Tutorials <https://github.com/NREL/bifacial_radiance/tree/master/docs/tutorials>`_.  
 * Source code for `gencumulativesky` is now included along with `makefile` instructions in ``/data/gencumsky/``
 * New Torquetube options added on py:class:`~bifacial_radiance.RadianceObj.makeModule`
 
@@ -18,8 +18,8 @@ API Changes
 * Input for daydate added to py:class:`~bifacial_radiance.RadianceObj.analysis1axis`, allowing to save cumulative files for daydate. 
 * Sorted trackerdict keys on simulations that use trackerdict routines
 * Removed deprecation warning and inputs ``nMods`` and ``nRows`` on py:class:`~bifacial_radiance.RadianceObj.makeScene1axis`
-* Modified ``quickExample`` to include testfolder as an input, and returned analysis results. Call to interactive directory selection called from `main.py` instead than being redefined inside the function. (:issue:`201`)
-* Removed ``HPCExample`` function from `main.py` (added as a Tutorial Notebook)
+* Modified `quickExample()` to include testfolder as an input, and returned analysis results. Call to interactive directory selection called from `main.py` instead than being redefined inside the function. (:issue:`201`)
+* Removed `HPCExample()` function from `main.py` (added as a `Tutorial Notebook <https://github.com/NREL/bifacial_radiance/tree/master/docs/tutorials>`_)
 * Fix GUI load issue when `None` values passed
 * Set GUI default ``modWanted`` and ``rowWanted`` to `Null`
 * py:class:`~bifacial_radiance.load.readconfigurationinputfile`.ini input files now use the key ``timeIndexSimulation`` instead of ``timestampRangeSimulation``
@@ -49,20 +49,20 @@ Testing
 
 Documentation
 ~~~~~~~~~~~~~~
-* Updated bifacial_radiance D.O.E. reference citation on :ref:`_index` (:issue:`213`)
-* New Tutorial on AgriPV example included in ''/docs/tutorials/12 AgriPV Systems.ipynb''
-* New Tutorial on running HPC Example
-* Added Webinar Slides and Tutorial Journal
-* Added `pip install` options to :ref:`_installation` 
-* Added software compatibility to :ref:`_installation` (:issue:`190`) 
-* Added Linux and MacOS instructions to :ref:`_installation` (:issue:`198`) 
-* Added content to :ref:`_package_overview` section (:issue:`191`)
+* Updated bifacial_radiance D.O.E. reference citation on index. (:issue:`213`)
+* New `Tutorial on AgriPV <https://github.com/NREL/bifacial_radiance/tree/master/docs/tutorials>`_
+* New `Tutorial on running High Performance Computing (HPC) <https://github.com/NREL/bifacial_radiance/tree/master/docs/tutorials>`_
+* Added `Webinar <https://www.youtube.com/watch?v=1X9L-R-RVGA>`_ Slides and Journal on `Tutorials <https://github.com.NREL/bifacial_radiance/tree/master/docs/tutorials>`_ 
+* Added `pip install` options to :ref:`installation` 
+* Added software compatibility to :ref:`installation` (:issue:`190`) 
+* Added Linux and MacOS instructions to :ref:`installation` (:issue:`198`) 
+* Added content to :ref:`package_overview` section (:issue:`191`)
[…]
