### 2020-04-17 4aeb935
remove development tutorial 10 - PVSyst comparison
