### 2020-04-17 3d736c1
setup.py drop py27, include pvmismatch
