### 2020-04-17 d83d2cc
No pandas 1.0 support yet
