+++ setup.py
@@ -97,6 +97,7 @@ setup(
     # requirements files see:
     # https://packaging.python.org/en/latest/requirements.html
     install_requires=[
+        'pandas < 1.0',
         'pvlib>= 0.6.1',
         'pytest',
         'pytest-cov',
