### 2020-04-17 f969a2b
pandas < 1.0. Update requirements.txt
