### 2020-04-24 0e38818
Check for pd.Series object in mad_fn. Add new pytests to check for it too. Fixes #240
