### 2020-04-24 ba13f0c
remove pandas < 1.0 requirement in setup.py
