### 2020-04-28 #1865 @usethedata
I've finally completed my review.  I found one minor documentation problem, which appears to be fixed in the repository.  I was able to successfully install and use the system on Mac OS Catalina.  Apologies for the delay, but the work from home situation means I have way more meetings, it seems.
