### 2020-04-28 46cd570
239 drop py27 (#241)
* setup.py drop py27, include pvmismatch

* pandas < 1.0. Update requirements.txt

* TravisCI build config update for py37 and upgrade-strategy=eager

* Check for pd.Series object in mad_fn. Add new pytests to check for it too. Fixes #240
