+++ requirements.txt
@@ -1,28 +1,26 @@
-atomicwrites==1.2.1
-attrs==18.2.0
-backports.functools-lru-cache==1.5
-colorama==0.4.1
-coverage==4.5.2
+atomicwrites==1.3.0
+attrs==19.3.0
+colorama==0.4.3
+configparser==5.0.0
+coverage==5.1
 cycler==0.10.0
-funcsigs==1.0.2
-ipython==5.8.0
-kiwisolver==1.0.1
-matplotlib==2.2.3
-more-itertools==5.0.0
-numpy==1.16.0
-pandas==0.24.0
-pathlib2==2.3.3
-pluggy==0.8.1
-pvlib==0.6.3
+idna==2.9
+importlib-metadata==1.6.0
+kiwisolver==1.2.0
+matplotlib==3.2.1
+more-itertools==8.2.0
+numpy==1.18.2
+pandas==0.25.3
+pluggy==0.13.1
+pvlib==0.7.1
 pvmismatch==4.1
-py==1.7.0
-pyparsing==2.3.1
-pytest==4.1.1
-pytest-cov==2.6.1
-python-dateutil==2.7.5
-pytz==2018.9
-scandir==1.9.0
-six==1.12.0
+py==1.8.1
+pyparsing==2.4.7
+pytest==5.4.1
+pytest-cov==2.8.1
+python-dateutil==2.8.1
+pytz==2019.3
+six==1.14.0
 sphinx-autoapi==1.1.0
 sphinx-rtd-theme==0.4.3
 requests
