### 2020-05-04 #1865 @melissawm
Since the review is complete, we can go forward with the review workflow. According to the [review guide](https://joss.readthedocs.io/en/latest/submitting.html#the-review-process),

> Upon successful completion of the review, authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as Zenodo or figshare, get a DOI for the archive, and update the review issue thread with the version number and DOI.

When you get a chance to do that, let me know so I can proofread the manuscript and we can proceed. Thanks!
