+++ bifacial_radiance/test2.py
@@ -1,3732 +1,0 @@
-#!/usr/bin/env python
-from __future__ import division  # avoid integer division issues.
-from __future__ import absolute_import # this module uses absolute imports
-"""
-@author: cdeline
-
-bifacial_radiance.py - module to develop radiance bifacial scenes, including 
-gendaylit and gencumulativesky.
-
-"""
-import logging
-logging.basicConfig()
-LOGGER = logging.getLogger(__name__)
-LOGGER.setLevel(logging.DEBUG)
-
-import os, datetime, sys
-from subprocess import Popen, PIPE  # replacement for os.system()
-import pandas as pd
-import numpy as np 
-from inputCostal1 import *
-
-# Mutual parameters across all processes
-#daydate=sys.argv[1]
-
-
-global DATA_PATH # path to data files including module.json.  Global context
-#DATA_PATH = os.path.abspath(pkg_resources.resource_filename('bifacial_radiance', 'data/') )
-DATA_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), 'data'))
-
-def _findme(lst, a): #find string match in a list. script from stackexchange
-    return [i for i, x in enumerate(lst) if x == a]
-
-
-def _normRGB(r, g, b): #normalize by each color for human vision sensitivity
-    return r*0.216+g*0.7152+b*0.0722
-
-def _popen(cmd, data_in, data_out=PIPE):
-    """
-    Helper function subprocess.popen replaces os.system
-    - gives better input/output process control
-    usage: pass <data_in> to process <cmd> and return results
-    based on rgbeimage.py (Thomas Bleicher 2010)
-    """
-    cmd = str(cmd) # gets rid of unicode oddities
-
-    p = Popen(cmd, bufsize=-1, stdin=PIPE, stdout=data_out, stderr=PIPE, shell=True) #shell=True required for Linux? quick fix, but may be security concern
-    data, err = p.communicate(data_in)
-    #if err:
-    #    return 'message: '+err.strip()
-    #if data:
-    #    return data. in Python3 this is returned as `bytes` and needs to be decoded
-    if err:
-        if data:
-            returntuple = (data.decode('latin1'), 'message: '+err.decode('latin1').strip())
-        else:
-            returntuple = (None, 'message: '+err.decode('latin1').strip())
-    else:
-        if data:
-            returntuple = (data.decode('latin1'), None) #Py3 requires decoding
-        else:
-            returntuple = (None, None)
-
-    return returntuple
-
-def _interactive_load(title=None):
-    # Tkinter file picker
-    import tkinter
-    from tkinter import filedialog
-    root = tkinter.Tk()
-    root.withdraw() #Start interactive file input
-    root.attributes("-topmost", True) #Bring window into foreground
-    return filedialog.askopenfilename(parent=root, title=title) #initialdir = data_dir
-
-def _interactive_directory(title=None):
-    # Tkinter directory picker.  Now Py3.6 compliant!
-    import tkinter
-    from tkinter import filedialog
-    root = tkinter.Tk()
-    root.withdraw() #Start interactive file input
-    root.attributes("-topmost", True) #Bring to front
-    return filedialog.askdirectory(parent=root, title=title)
-
-
-class RadianceObj:
-    """
-    The RadianceObj top level class is used to work on radiance objects, 
-    keep track of filenames,  sky values, PV module configuration, etc.
-
-    Parameters
-    ----------
-    name : text to append to output files
-    filelist : list of Radiance files to create oconv
-    nowstr : current date/time string
-    path : working directory with Radiance materials and objects
-
-    Methods
-    -------
-    __init__ : initialize the object
-    _setPath : change the working directory
-
-    """
-
-    def __init__(self, name=None, path=None):
-        '''
[…]
