### 2020-05-07 7b7b8d1
allow pvlib > 0.6.1
