+++ requirements.txt
@@ -1,5 +1,5 @@
 atomicwrites==1.3.0
-attrs==19.3.0
+attrs==19.1.0
 colorama==0.4.3
 configparser==5.0.0
 coverage==5.1
