### 2020-05-07 97ed977
return MetObj.albedo from TMY file
