+++ bifacial_radiance/main.py
@@ -2300,84 +2300,70 @@ class GroundObj:
 
         self.normval = ''
         self.ReflAvg = ''
-        self.ground_type = ''
         self.material_options = []
         self.Rrefl = ''
         self.Grefl = ''
         self.Brefl = ''
-        albedo = None
         material_path = 'materials'
+        self.ground_type = 'custom'        
 
         if material_file is None:
             material_file = 'ground.rad'
 
-        #check if materialOrAlbedo is a float between 0 and 1
-        try:
-            albedo = float(materialOrAlbedo)
-            if not (0 < albedo < 1):
-                albedo = None
-                materialOrAlbedo = None
-            else:
-                albedo = [albedo, albedo, albedo]
-        except TypeError:
-            # nothing passed
-            try: 
-                if len(materialOrAlbedo) == 3:
-                    if not ((0 < float(materialOrAlbedo[0]) < 1) and 
-                            (0 < float(materialOrAlbedo[1]) < 1) and 
-                            (0 < float(materialOrAlbedo[2]) < 1)) :
-                        print("Reflectivity values must be between 0 and 1")
-                        albedo = None
-                    else:
-                        albedo = materialOrAlbedo
-                else:
-                    print("Wrong albedo type passed. Either a single value or a ",
-                          "list of 3 values expected (0.62 or [0.62, 0.62, 0.62] for each wavelength")
-            except:
-                  # nothing passed
-                  albedo = None
-        except ValueError:
-            # material string passed
-            albedo = None
+        if type(materialOrAlbedo) is str:
+            self.ground_type = materialOrAlbedo  
 
-        if albedo is not None:
-            self.Rrefl = albedo[0]
-            self.Grefl = albedo[1]
-            self.Brefl = albedo[2]
-            self.normval = _normRGB(albedo[0],albedo[1],albedo[2])
-            self.ReflAvg = np.round(np.mean(albedo),4)
-            self.ground_type = 'custom'
-
-        else:
             f = open(os.path.join(material_path,material_file))
+            #f = open(r'C:\Users\sayala\Documents\GitHub\bifacial_radiance\bifacial_radiance\TEMP\materials\ground.rad')
             keys = [] #list of material key names
-            Rrefl = []; Grefl=[]; Brefl=[] #RGB reflectance of the material
+            Rreflall = []; Greflall=[]; Breflall=[] #RGB reflectance of the material
             temp = f.read().split()
             f.close()
             #return indices for 'plastic' definition
             index = _findme(temp,'plastic')
             for i in index:
                 keys.append(temp[i+1])# after plastic comes the material name
-                Rrefl.append(float(temp[i+5]))#RGB reflectance comes a few more down the list
-                Grefl.append(float(temp[i+6]))
-                Brefl.append(float(temp[i+7]))
-
+                Rreflall.append(float(temp[i+5]))#RGB reflectance comes a few more down the list
+                Greflall.append(float(temp[i+6]))
+                Breflall.append(float(temp[i+7]))
+        
             self.material_options = keys
-
+        
             if materialOrAlbedo is not None:
                 # if material isn't specified, return list of material options
                 index = _findme(keys,materialOrAlbedo)[0]
-                # calculate avg reflectance of the material and normalized color
-                # using NormRGB function
-                self.normval = _normRGB(Rrefl[index],Grefl[index],Brefl[index])
-                self.ReflAvg = (Rrefl[index]+Grefl[index]+Brefl[index])/3
-                self.ground_type = keys[index]
[…]
