### 2020-05-08 71a5f19
setGround: attempt to read metdata and assign albedo; handling of lists
setGround: attempt to read metdata and assign albedo; handling of lists
