### 2020-05-08 1df437c
gencumsky albedo averaging
