### 2020-05-08 e2df2cf
Undo gendaylit API change to get tests passing.  possibly change at a later time.
