### 2020-05-08 1030499
Initialize groundobj testing - works on Master cases
