### 2020-05-09 3a99608
warn on > 3 albedo values and trim down to 3.  Tests passing
