### 2020-05-09 b8a21f7
Change gendaylit API - metdata is optional parameter in second position.  ALL notebooks and tests updated
