### 2020-05-11 4efbc11
Create printGroundplane() function in groundobj. Refactor gendaylit, gencumsky
