+++ bifacial_radiance/main.py
@@ -387,7 +387,7 @@ class RadianceObj:
         self.ground : tuple
             self.ground.normval : numeric
             Normalized color value
-            self.gorund.ReflAvg : numeric
+            self.ground.ReflAvg : numeric
             Average reflectance
         """
 
@@ -787,10 +787,12 @@ class RadianceObj:
                       'readWeatherfile(), readEPW() or readTMY()') 
                 return
 
-        if type(timeindex)== MetObj:
+        if type(timeindex)== MetObj:  # check for deprecated usage of gendaylit
             warnings.warn('passed MetObj into timeindex position - proper ' +
                           'usage: gendaylit(timeindex, metdata) ')
             return
+        
+        ground = self.ground
         
         locName = metdata.city
         dni = metdata.dni[timeindex]
@@ -802,29 +804,19 @@ class RadianceObj:
 
         # Assign Albedos
         try:
-            if self.ground.ReflAvg.shape == metdata.dni.shape:
-                ReflAvg = self.ground.ReflAvg[timeindex]
-                Rrefl = self.ground.Rrefl[timeindex]
-                Grefl = self.ground.Grefl[timeindex]
-                Brefl =  self.ground.Brefl[timeindex]
-                normval = self.ground.normval[timeindex]        
+            if ground.ReflAvg.shape == metdata.dni.shape:
+                groundindex = timeindex  
             elif self.ground.ReflAvg.shape[0] == 1: # just 1 entry
-                ReflAvg = self.ground.ReflAvg[0]
-                Rrefl = self.ground.Rrefl[0]
-                Grefl = self.ground.Grefl[0]
-                Brefl =  self.ground.Brefl[0]
-                normval = self.ground.normval[0]
+                groundindex = 0
             else:
                 print("Shape of ground Albedos and TMY data do not match.")
                 return
         except:
             print('usage: make sure to run setGround() before gendaylit()')
             return
-                
-        ground_type = self.ground.ground_type
 
         if debug is True:
-            print('Sky generated with Gendaylit 2, with DNI: %0.1f, DHI: %0.1f' % (dni, dhi))
+            print('Sky generated with Gendaylit, with DNI: %0.1f, DHI: %0.1f' % (dni, dhi))
             print("Datetime TimeIndex", metdata.datetime[timeindex])
 
 
@@ -859,18 +851,10 @@ class RadianceObj:
             +" LON: " + str(lon) + " Elev: " + str(elev) + "\n"
             "# Sun position calculated w. PVLib\n" + \
             "!gendaylit -ang %s %s" %(sunalt, sunaz)) + \
-            " -W %s %s -g %s -O 1 \n" %(dni, dhi, ReflAvg) + \
+            " -W %s %s -g %s -O 1 \n" %(dni, dhi, ground.ReflAvg[groundindex]) + \
             "skyfunc glow sky_mat\n0\n0\n4 1 1 1 0\n" + \
             "\nsky_mat source sky\n0\n0\n4 0 0 1 180\n" + \
-            '\nskyfunc glow ground_glow\n0\n0\n4 ' + \
-            '%s ' % (Rrefl/normval)  + \
-            '%s ' % (Grefl/normval) + \
-            '%s 0\n' % (Brefl/normval) + \
-            '\nground_glow source ground\n0\n0\n4 0 0 -1 180\n' +\
-            "\nvoid plastic %s\n0\n0\n5 %0.3f %0.3f %0.3f 0 0\n" %(
-            ground_type, Rrefl, Grefl, Brefl) +\
-            "\n%s ring groundplane\n" % (ground_type) +\
-            '0\n0\n8\n0 0 -.01\n0 0 1\n0 100'
[…]
