### 2020-05-11 13a06bc
Various cleanup
