+++ bifacial_radiance/main.py
@@ -809,7 +809,7 @@ class RadianceObj:
             elif self.ground.ReflAvg.shape[0] == 1: # just 1 entry
                 groundindex = 0
             else:
-                print("Shape of ground Albedos and TMY data do not match.")
+                warnings.warn("Shape of ground Albedos and TMY data do not match.")
                 return
         except:
             print('usage: make sure to run setGround() before gendaylit()')
@@ -2335,14 +2335,19 @@ class GroundObj:
 
         if material_file is None:
             material_file = 'ground.rad'
+        
+        if materialOrAlbedo is None: # Case where it's none.
+                print('\nInput albedo 0-1, or ground.rad name e.g. litesoil')
+                print('\nAlternatively, run setGround after readWeatherData()',
+                      ' and setGround will read metdata.albedo if availalbe')
+                return
             
-        if type(materialOrAlbedo) is str or materialOrAlbedo is None:
+        if type(materialOrAlbedo) is str :
             self.ground_type = materialOrAlbedo  
 
             f = open(os.path.join(material_path,material_file))
-            #f = open(r'C:\Users\sayala\Documents\GitHub\bifacial_radiance\bifacial_radiance\TEMP\materials\ground.rad')
             keys = [] #list of material key names
-            Rreflall = []; Greflall=[]; Breflall=[] #RGB reflectance of the material
+            Rreflall = []; Greflall=[]; Breflall=[] #RGB material reflectance  
             temp = f.read().split()
             f.close()
             #return indices for 'plastic' definition
@@ -2354,19 +2359,14 @@ class GroundObj:
                 Breflall.append(float(temp[i+7]))
         
             self.material_options = keys
-        
-            if materialOrAlbedo is not None:
-                # if material isn't specified, return list of material options
-                index = _findme(keys,materialOrAlbedo)[0]
-                materialOrAlbedo = np.array([[Rreflall[index], Greflall[index], Breflall[index]]])
-            else: # Case wehre it's none.
-                print('\nInput albedo 0-1, or ground material names:'+str(keys))
-                print('\nAlternatively, run setGround after readWeatherData function',
-                      'and setGround will read metdata.albedo if availalbe')
-                return
+
+            index = _findme(keys,materialOrAlbedo)[0]
+            materialOrAlbedo = np.array([[Rreflall[index], Greflall[index], Breflall[index]]])
+
             
         if type(materialOrAlbedo) is float:
-            materialOrAlbedo = np.array([[materialOrAlbedo, materialOrAlbedo, materialOrAlbedo]])
+            materialOrAlbedo = np.array([[materialOrAlbedo, 
+                                          materialOrAlbedo, materialOrAlbedo]])
         
         if type(materialOrAlbedo) is list:
             materialOrAlbedo = np.asarray(materialOrAlbedo)
@@ -2375,23 +2375,29 @@ class GroundObj:
         if type(materialOrAlbedo) is np.ndarray:
 
             if materialOrAlbedo.ndim == 0:
-            # If a numpy array of one signle value was passed, i.e. np.array(0.62)
+            # numpy array of one single value, i.e. np.array(0.62)
             # after this if, np.array([0.62])
                 materialOrAlbedo = materialOrAlbedo.reshape([1])
                 
             if materialOrAlbedo.ndim == 1:
             # If np.array is ([0.62]), this repeats it so at the end it's
             # np.array ([0.62, 0.62, 0.62])
-                materialOrAlbedo = np.repeat(np.array([materialOrAlbedo]), 3, axis=1).reshape(len(materialOrAlbedo),3)
+                materialOrAlbedo = np.repeat(np.array([materialOrAlbedo]), 
+                                             3, axis=1).reshape(
[…]
