+++ bifacial_radiance/mismatch.py
@@ -40,7 +40,8 @@ def _sensorupsampletocellsbyInterpolation(df, cellsy):
     
 def _sensorsdownsampletocellsbyAverage(df, cellsy):
     '''
-    df = dataframe with rows indexed by number (i.e. 0 to sensorsy) where sensorsy > cellsy
+    df = dataframe with rows indexed by number (i.e. 0 to sensorsy) 
+         where sensorsy > cellsy
     cellsy = int. usually 8 or 12.
 
     example:
@@ -52,7 +53,8 @@ def _sensorsdownsampletocellsbyAverage(df, cellsy):
     edges=len(df)-np.floor(len(df)/(cellsy))*(cellsy)
     edge1=int(np.floor(edges/2))
     edge2=int(edges-edge1)
-    A = list(range(df.index[0]+edge1, df.index[-1]-edge2+2, int(np.floor(len(df)/(cellsy)))))
+    A = list(range(df.index[0]+edge1, df.index[-1]-edge2+2, 
+                   int(np.floor(len(df)/(cellsy)))))
     B = range(0,len(A)-1,1)
     C = [df.iloc[A[x]:A[x+1]].mean(axis=0) for x in B]
     df_centeraverages=pd.DataFrame(C)
@@ -61,7 +63,8 @@ def _sensorsdownsampletocellsbyAverage(df, cellsy):
 
 def _sensorsdownsampletocellbyCenter(df, cellsy):
     '''
-    df = dataframe with rows indexed by number (i.e. 0 to sensorsy) where sensorsy > cellsy
+    df = dataframe with rows indexed by number (i.e. 0 to sensorsy) 
+         where sensorsy > cellsy
     cellsy = int. usually 8 or 12.
 
     example:
@@ -74,7 +77,8 @@ def _sensorsdownsampletocellbyCenter(df, cellsy):
     edges=len(df)-np.floor(len(df)/(cellsy))*(cellsy)
     edge1=int(np.floor(edges/2))
     edge2=int(edges-edge1)
-    A = list(range(df.index[0]+edge1, df.index[-1]-edge2+2, int(np.floor(len(df)/(cellsy)))))
+    A = list(range(df.index[0]+edge1, df.index[-1]-edge2+2, 
+                   int(np.floor(len(df)/(cellsy)))))
     A = [int(x+(A[1]-A[0])*0.5) for x in A]
     A = A[:-1]
     df_centervalues=df.loc[A]
@@ -86,7 +90,8 @@ def _setupforPVMismatch(portraitorlandscape, sensorsy, numcells=72):
     r''' Sets values for calling PVMismatch, for ladscape or portrait modes and 
     
     Example:
-    stdpl, cellsx, cellsy = _setupforPVMismatch(portraitorlandscape='portrait', sensorsy=100):
+    stdpl, cellsx, cellsy = _setupforPVMismatch(portraitorlandscape='portrait', 
+                                                sensorsy=100):
     '''
 
     import numpy as np
@@ -120,13 +125,15 @@ def _setupforPVMismatch(portraitorlandscape, sensorsy, numcells=72):
             [10,	13,	34,	37,	58,	61,	82,	85],
             [11,	12,	35,	36,	59,	60,	83,	84]])
     else:
-        print("Error. Only 72 and 96 cells modules supported at the moment. Change numcells to either of this options!")
+        print("Error. Only 72 and 96 cells modules supported at the moment. "
+              "Change numcells to either of this options!")
         return
     
     if portraitorlandscape == 'landscape':
         stdpl = stdpl.transpose()
     elif portraitorlandscape != 'portrait':
-        print("Error. portraitorlandscape variable must either be 'landscape' or 'portrait'")
[…]
