### 2020-05-12 7e45058
create printGroundMaterials() function in GroundObj
