### 2020-05-12 ac27ce1
Add testing. comment out load_inputvariablesfile because it won't work in Python3
