### 2020-05-12 a0a35ff
Issue68 hourly albedo (#246)
* GroundObj for Arrays.  Works for all cases of albedos:
materialoralbedo1 = np.array(0.62) # 1 albedo
materialoralbedo2 = np.array([0.62]) # 1 Albedo
materialoralbedo3 = np.array([0.52, 0.33, 0.44, 0.55]) # 4 Albedoss
materialoralbedo4 = np.array([[0.2, 0.3, 0.4]]) # 1 RGB Albedo
materialoralbedo5 = np.array([[0.2, 0.3, 0.4],  [0.12, 0.13, 0.26]]) # 2 RGB Albedo

* setGround: attempt to read metdata and assign albedo; handling of lists

* return MetObj.albedo from TMY file

* gendaylit Update for Albedo.  Changes: 1) Made metdata input optional, as it is saved in 
[…]
