### 2020-05-18 79d4141
Pep8 cleanup (#248)
* AutoPep8 cleanup
