### 2020-05-18 7012fbb
Revert "Merge branch '0333_release' of https://github.com/NREL/bifacial_radiance into 0333_release"
This reverts commit 5cff331081ac7dfe4d5dfe200674b3d491a7f24e, reversing
changes made to 2f1e4ad1c981e5a5548f2d06b84fb84e2dd9f4b3.
