+++ bifacial_radiance/main.py
@@ -62,21 +62,19 @@ import numpy as np
 #from input import *
 
 # Mutual parameters across all processes
-# daydate=sys.argv[1]
+#daydate=sys.argv[1]
 
 
-global DATA_PATH  # path to data files including module.json.  Global context
+global DATA_PATH # path to data files including module.json.  Global context
 #DATA_PATH = os.path.abspath(pkg_resources.resource_filename('bifacial_radiance', 'data/') )
 DATA_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__), 'data'))
 
-
-def _findme(lst, a):  # find string match in a list. script from stackexchange
+def _findme(lst, a): #find string match in a list. script from stackexchange
     return [i for i, x in enumerate(lst) if x == a]
 
 
-def _normRGB(r, g, b):  # normalize by each color for human vision sensitivity
+def _normRGB(r, g, b): #normalize by each color for human vision sensitivity
     return r*0.216+g*0.7152+b*0.0722
-
 
 def _popen(cmd, data_in, data_out=PIPE):
     """
@@ -85,49 +83,43 @@ def _popen(cmd, data_in, data_out=PIPE):
     usage: pass <data_in> to process <cmd> and return results
     based on rgbeimage.py (Thomas Bleicher 2010)
     """
-    cmd = str(cmd)  # gets rid of unicode oddities
+    cmd = str(cmd) # gets rid of unicode oddities
 
-    # shell=True required for Linux? quick fix, but may be security concern
-    p = Popen(cmd, bufsize=-1, stdin=PIPE,
-              stdout=data_out, stderr=PIPE, shell=True)
+    p = Popen(cmd, bufsize=-1, stdin=PIPE, stdout=data_out, stderr=PIPE, shell=True) #shell=True required for Linux? quick fix, but may be security concern
     data, err = p.communicate(data_in)
-    # if err:
+    #if err:
     #    return 'message: '+err.strip()
-    # if data:
+    #if data:
     #    return data. in Python3 this is returned as `bytes` and needs to be decoded
     if err:
         if data:
-            returntuple = (data.decode('latin1'), 'message: ' +
-                           err.decode('latin1').strip())
+            returntuple = (data.decode('latin1'), 'message: '+err.decode('latin1').strip())
         else:
             returntuple = (None, 'message: '+err.decode('latin1').strip())
     else:
         if data:
-            # Py3 requires decoding
-            returntuple = (data.decode('latin1'), None)
+            returntuple = (data.decode('latin1'), None) #Py3 requires decoding
         else:
             returntuple = (None, None)
 
     return returntuple
-
 
 def _interactive_load(title=None):
     # Tkinter file picker
     import tkinter
     from tkinter import filedialog
     root = tkinter.Tk()
-    root.withdraw()  # Start interactive file input
-    root.attributes("-topmost", True)  # Bring window into foreground
-    return filedialog.askopenfilename(parent=root, title=title)
-
+    root.withdraw() #Start interactive file input
+    root.attributes("-topmost", True) #Bring window into foreground
+    return filedialog.askopenfilename(parent=root, title=title) #initialdir = data_dir
 
 def _interactive_directory(title=None):
     # Tkinter directory picker.  Now Py3.6 compliant!
     import tkinter
     from tkinter import filedialog
     root = tkinter.Tk()
-    root.withdraw()  # Start interactive file input
-    root.attributes("-topmost", True)  # Bring to front
+    root.withdraw() #Start interactive file input
+    root.attributes("-topmost", True) #Bring to front
     return filedialog.askdirectory(parent=root, title=title)
 
 
@@ -169,11 +161,11 @@ class RadianceObj:
         self.data = {}           # data stored at each timestep
[…]
