+++ bifacial_radiance/mismatch.py
@@ -5,43 +5,42 @@ Created on Tue Mar 26 20:16:47 2019
 @author: sayala
 """
 
-#from load import *
+#from load import * 
+
 
 
 def _sensorupsampletocellsbyInterpolation(df, cellsy):
     '''
-
+        
     Function for when sensorsy in the results are less than cellsy desired.
     Interpolates the dataframe.
-
+    
     #2DO: improve interpolation with pandas. right onw it's row by row.
-
+    
     _sensorupsampletocellsbyInterpolation(df, cellsy)
     '''
-
+    
     import pandas as pd
     import numpy as np
-
+    
     sensorsy = len(df)
-
-    # 2DO: Update this section to match bifacialvf
-    cellCenterPVM = []
-    for i in range(0, cellsy):
+    
+    #2DO: Update this section to match bifacialvf
+    cellCenterPVM=[]                                    
+    for i in range (0, cellsy):
         cellCenterPVM.append((i*sensorsy/cellsy+(i+1)*sensorsy/cellsy)/2)
-
+    
     df2 = pd.DataFrame()
-    for j in range(0, len(df.keys())):
+    for j in range (0, len(df.keys())):
         A = list(df[df.keys()[j]])
-        B = np.interp(cellCenterPVM, list(range(0, sensorsy)), A)
-        df2[df.keys()[j]] = B
-
+        B= np.interp(cellCenterPVM, list(range(0,sensorsy)), A)
+        df2[df.keys()[j]]=B
+    
     return df2
-
-
+    
 def _sensorsdownsampletocellsbyAverage(df, cellsy):
     '''
-    df = dataframe with rows indexed by number (i.e. 0 to sensorsy)
-         where sensorsy > cellsy
+    df = dataframe with rows indexed by number (i.e. 0 to sensorsy) where sensorsy > cellsy
     cellsy = int. usually 8 or 12.
 
     example:
@@ -50,22 +49,19 @@ def _sensorsdownsampletocellsbyAverage(df, cellsy):
     import numpy as np
     import pandas as pd
 
-    edges = len(df)-np.floor(len(df)/(cellsy))*(cellsy)
-    edge1 = int(np.floor(edges/2))
-    edge2 = int(edges-edge1)
-    A = list(range(df.index[0]+edge1, df.index[-1]-edge2+2,
-                   int(np.floor(len(df)/(cellsy)))))
-    B = range(0, len(A)-1, 1)
+    edges=len(df)-np.floor(len(df)/(cellsy))*(cellsy)
+    edge1=int(np.floor(edges/2))
+    edge2=int(edges-edge1)
+    A = list(range(df.index[0]+edge1, df.index[-1]-edge2+2, int(np.floor(len(df)/(cellsy)))))
+    B = range(0,len(A)-1,1)
     C = [df.iloc[A[x]:A[x+1]].mean(axis=0) for x in B]
-    df_centeraverages = pd.DataFrame(C)
-
+    df_centeraverages=pd.DataFrame(C)
+    
     return df_centeraverages
-
 
 def _sensorsdownsampletocellbyCenter(df, cellsy):
     '''
-    df = dataframe with rows indexed by number (i.e. 0 to sensorsy)
-         where sensorsy > cellsy
+    df = dataframe with rows indexed by number (i.e. 0 to sensorsy) where sensorsy > cellsy
     cellsy = int. usually 8 or 12.
 
     example:
@@ -74,110 +70,103 @@ def _sensorsdownsampletocellbyCenter(df, cellsy):
 
     import numpy as np
 
-    edges = len(df)-np.floor(len(df)/(cellsy))*(cellsy)
[…]
