### 2020-05-20 #1865 @arfon
👋 @shirubana - have you managed to complete the steps that @melissawm asked of you a few weeks ago? JOSS is now open again for submissions and we're actively trying to wrap up existing reviews.
