### 2020-05-20 24d91e9
AttributeError - bug fix in analysis1axis line 2207.  Fixes #251
