### 2020-05-20 abef1ff
Allow int's to be passed to setGround(), lump in with floats.
