### 2020-05-20 fcb1339
Use nonzeromean() to exclude zero values from albedo averages with gencumsky.  Account for zero albedo in _makeGroundString.  Fixes #251
