### 2020-05-20 bb3148e
Move _nonzeromean up to root of GroundObj. Use it for albedo load message
