### 2020-05-20 047a785
update to gendaylit documentation for MM_DD_HH
