### 2020-05-20 971eab1
ENH: Enable module.json human readability (#231)
* use 'python -m json.tool --sort-keys' to pretty print & sort module.json

* indent and sort keys for module.json when writing new modules

Co-authored-by: cdeline <chris.deline@nrel.gov>
