### 2020-05-20 5acda4e
hpc flag irrplot #209
addresses issue 209 where irrplot needs hpc flag but calling function analysis1axis does not provide it.
