### 2020-05-21 015e80f
genCumSky1axis bug fix for partial year datafile. Fixes #203
