### 2020-05-21 2070fbe
Constrain albedo to 0:1.  Fixes #227
