### 2020-05-21 40fa765
GUI tracker visualization enabled. Fixes #229, zero tilt only.
