### 2020-05-21 9d77259
makeModule informational messages updated.  Fixes #202
