### 2020-05-21 15ebff2
whatsnew update. remove if __name__ = main run option from main.py
