+++ docs/sphinx/source/whatsnew/v0.3.3.3.rst
@@ -1,36 +1,0 @@
-.. _whatsnew_0333:
-
-v0.3.3.3 (X / X / 2020)
---------------------------
-Release of new version including RGB and time-series albedo options 
-
-
-Enhancements
-~~~~~~~~~~~~
-* 
-API Changes
-~~~~~~~~~~~
-* Input changed to py:class:`~bifacial_radiance.RadianceObj.gendaylit`, now making metdata an optional input. 
-
-
-Bug fixes
-~~~~~~~~~
-* 
-
-Testing
-~~~~~~~
-* Added pytesting to py:class:`~bifacial_radiance.GroundObj`
-
-
-Documentation
-~~~~~~~~~~~~~~
-
-
-Contributors
-~~~~~~~~~~~~
-* Chris Deline (:ghuser:`cdeline`)
-* Silvana Ayala (:ghuser:`shirubana`)
-
-
-
-
