### 2020-05-21 01d0a44
revert if __name__ == "__main__" removal. keep in development.
