### 2020-05-21 9705acd
Jupyter fixes IMAGES calls
Some of the images are not showing because they are labeled '.png' instead of 'PNG'. Modified
