### 2020-05-21 ed1950e
image corrections and some advanced journals paths
two of the journals had not a multiple OS friendly path so updated #217

Fixed images.
