### 2020-05-22 013ad54
Added Journal 13 and delete development journals
Deleted UV validation journal
Deleted Sensor position since the error in orientation was fixed in last release
Added Glass Module updated journal
Updated some other details in other journals.
