+++ docs/tutorials/7 - Advanced topics - Multiple SceneObjects Example.py
@@ -12,7 +12,7 @@
 # 
 # A scene Object is defined as an array of modules, with whatever parameters you want to give it. In this case, we are modeling one array of 2 rows of 5 modules in landscape, and one array of 1 row of 5 modules in 2-UP, portrait configuration, as the image below:
 # 
-# ![multiple Scene Objects Example](..\images_wiki\Journal_example_multiple_objects.PNG)
+# ![multiple Scene Objects Example](../images_wiki/Journal_example_multiple_objects.PNG)
 # 
 # 
 # ### Steps:
@@ -153,7 +153,7 @@ octfile = demo.makeOct(demo.getfilelist())
 # 
 # It should look something like this:
 # 
-# ![multiple Scene Objects Example](..\images_wiki\Journal_example_multiple_objects.PNG)
+# ![multiple Scene Objects Example](../images_wiki/Journal_example_multiple_objects.PNG)
 # 
 
 # <a id='step4'></a>
