### 2020-05-22 cdc85b8
whatsnew documentation fixes
in many places in documentation it said py:class:`~bifacial_radiance... instead of :py:class
