+++ docs/sphinx/source/whatsnew/v0.2.1.rst
@@ -1,6 +1,6 @@
 .. _whatsnew_0201:
 
-v0.2.1 (?? / ?? / 2019)
+v0.2.1 (05 / 21 / 2018)
 ------------------------
 
 * Allow TMY3 input files
