+++ docs/sphinx/source/whatsnew/v0.0.1.rst
@@ -1,6 +1,6 @@
 .. _whatsnew_0001:
 
-v0.0.1 (?? / ?? / 2018)
+v0.0.1 (05 / 24 / 2017)
 ------------------------
 
 Initial stable release.
