### 2020-05-22 647ca60
labelstyle added to MetObj.
default = left, like TMY

still need to run tests for all cases
