+++ bifacial_radiance/main.py
@@ -592,16 +592,27 @@ class RadianceObj:
         return tmydata_trunc
         
         
-    def readTMY(self, tmyfile=None, starttime=None, endtime=None, daydate=None):
+    def readTMY(self, tmyfile=None, starttime=None, endtime=None, daydate=None, labelstyle = 'left'):
         '''
         use pvlib to read in a tmy3 file.
 
         Parameters
         ------------
-        tmyfile:  filename of tmy3 to be read with pvlib.tmy.readtmy3
-        starttime:  'MM_DD_HH' string for limited time temp file
-        endtime:  'MM_DD_HH' string for limited time temp file
-        daydate : str for single day in 'MM/DD' or MM_DD format.
+        tmyfile : str
+            Filename of tmy3 to be read with pvlib.tmy.readtmy3
+        starttime : str 
+            'MM_DD_HH' string for limited time temp file
+        endtime: str
+            'MM_DD_HH' string for limited time temp file
+        daydate : str 
+            For single day in 'MM/DD' or MM_DD format.
+        labelstyle : str
+            'left', 'right', or 'exact'. For data that is averaged, defines if
+            the average is closed to the left, to hte right, or if exact timestamp
+            should be used for calculating sunposition. For example, TMY3 data
+            is closed to the left, so 11 AM data represents data from 10 to 11,
+            and sun position should be calculated at 10:30 AM.
+            Currently, SAM and PVSyst are using right-labeled for TMY data.        
         
         Returns
         -------
@@ -626,24 +637,35 @@ class RadianceObj:
         if daydate is not None:  # also remove GHI = 0 for HPC daydate call.
             tmydata_trunc = tmydata_trunc[tmydata_trunc.GHI > 0]
             
-        self.metdata = MetObj(tmydata_trunc, metadata)
+        self.metdata = MetObj(tmydata_trunc, metadata, labelstyle = labelstyle)
         return self.metdata
 
-    def readEPW(self, epwfile=None, hpc=False, starttime=None, endtime=None, daydate=None):
+    def readEPW(self, epwfile=None, hpc=False, starttime=None, endtime=None, daydate=None, labelstyle = 'left'):
         """
         Uses readepw from pvlib>0.6.1 but un-do -1hr offset and
         rename columns to match TMY3: DNI, DHI, GHI, DryBulb, Wspd
-        
+    
         Parameters
-        ----------
+        ------------
         epwfile : str
-            Direction and filename of the epwfile. 
-            If None, opens interactive loading window.
+            Direction and filename of the epwfile. If None, opens an interactive
+            loading window.
         hpc : bool
             Default False.  DEPRECATED
-        daydate : str for single day in 'MM/DD' or MM_DD format.  
-        starttime:  'MM_DD_HH' string for limited time temp file
-        endtime:  'MM_DD_HH' string for limited time temp file
+        starttime : str 
+            'MM_DD_HH' string for limited time temp file
+        endtime: str
+            'MM_DD_HH' string for limited time temp file
+        daydate : str 
+            For single day in 'MM/DD' or MM_DD format.
+        labelstyle : str
+            'left', 'right', or 'exact'. For data that is averaged, defines if
+            the average is closed to the left, to hte right, or if exact timestamp
+            should be used for calculating sunposition. For example, TMY3 data
+            is closed to the left, so 11 AM data represents data from 10 to 11,
+            and sun position should be calculated at 10:30 AM.
[…]
