### 2020-05-22 d81c9c5
corrected sunposition for sunrise and susnet for left and right labeled
For LEFT Labeled
HOUR --> SUPOSITION
11:00  --> 10:30
Sunrise at 7:21,
8 AM --> 7:41
Sunset at 16:46,
17:00 PM --> 16:23

For RIGHT Labeled
HOUR --> SUPOSITION
11 --> 11:30
Sunrise at 7:21,
7 AM --> 7:40
Sunset at 16:46,
16 --> 16:23
