### 2020-05-26 2c145ad
sky and module .rad name improvements
trying to add clarity
