### 2020-05-27 29a7b78
Change labelstyle to 'label'; change 'exact' to 'center'; add pytests for right vs left label. Passing.
