+++ bifacial_radiance/main.py
@@ -585,14 +585,16 @@ class RadianceObj:
             savedata = savedata.resample('1h').asfreq(fill_value=0)
         csvfile = os.path.join('EPWs', filename)
         print('Saving file {}, # points: {}'.format(csvfile, savedata.__len__()))
-        savedata.to_csv(csvfile, index=False, header=False, sep=' ', columns=['GHI','DHI'])
+        savedata.to_csv(csvfile, index=False, header=False, sep=' ', 
+                        columns=['GHI','DHI'])
         self.epwfile = csvfile
         
         # return tmydata truncated by startdt and enddt
         return tmydata_trunc
         
         
-    def readTMY(self, tmyfile=None, starttime=None, endtime=None, daydate=None, labelstyle = 'left'):
+    def readTMY(self, tmyfile=None, starttime=None, endtime=None, daydate=None, 
+                label = 'right'):
         '''
         use pvlib to read in a tmy3 file.
 
@@ -606,13 +608,14 @@ class RadianceObj:
             'MM_DD_HH' string for limited time temp file
         daydate : str 
             For single day in 'MM/DD' or MM_DD format.
-        labelstyle : str
-            'left', 'right', or 'exact'. For data that is averaged, defines if
-            the average is closed to the left, to hte right, or if exact timestamp
-            should be used for calculating sunposition. For example, TMY3 data
-            is closed to the left, so 11 AM data represents data from 10 to 11,
-            and sun position should be calculated at 10:30 AM.
-            Currently, SAM and PVSyst are using right-labeled for TMY data.        
+        label : str
+            'left', 'right', or 'center'. For data that is averaged, defines if
+            the timestamp refers to the left edge, the right edge, or the 
+            center of the averaging interval, for purposes of calculating 
+            sunposition. For example, TMY3 data is right-labeled, so 11 AM data 
+            represents data from 10 to 11, and sun position is calculated 
+            at 10:30 AM.  Currently SAM and PVSyst use left-labeled interval 
+            data and NSRDB uses centered.
         
         Returns
         -------
@@ -637,10 +640,11 @@ class RadianceObj:
         if daydate is not None:  # also remove GHI = 0 for HPC daydate call.
             tmydata_trunc = tmydata_trunc[tmydata_trunc.GHI > 0]
             
-        self.metdata = MetObj(tmydata_trunc, metadata, labelstyle = labelstyle)
+        self.metdata = MetObj(tmydata_trunc, metadata, label = label)
         return self.metdata
 
-    def readEPW(self, epwfile=None, hpc=False, starttime=None, endtime=None, daydate=None, labelstyle = 'left'):
+    def readEPW(self, epwfile=None, hpc=False, starttime=None, endtime=None, 
+                daydate=None, label = 'right'):
         """
         Uses readepw from pvlib>0.6.1 but un-do -1hr offset and
         rename columns to match TMY3: DNI, DHI, GHI, DryBulb, Wspd
@@ -658,13 +662,14 @@ class RadianceObj:
             'MM_DD_HH' string for limited time temp file
         daydate : str 
             For single day in 'MM/DD' or MM_DD format.
-        labelstyle : str
-            'left', 'right', or 'exact'. For data that is averaged, defines if
-            the average is closed to the left, to hte right, or if exact timestamp
-            should be used for calculating sunposition. For example, TMY3 data
-            is closed to the left, so 11 AM data represents data from 10 to 11,
-            and sun position should be calculated at 10:30 AM.
-            Currently, SAM and PVSyst are using right-labeled for TMY data.      
+        label : str
[…]
