### 2020-05-27 29737c5
comment out quickExample pytest.
