### 2020-05-27 c360d40
propagating label to readWeatherFile
