### 2020-05-27 3fc186b
whatsnew v 0.3.4 updated (and defaultreadWeatherFile = 'right')
