### 2020-05-27 59be383
update v0.3.4.rst
@cdeline I'm happy with the documentation now :) read to PR and PR into release? :)
