+++ docs/sphinx/source/whatsnew/v0.3.4.rst
@@ -2,39 +2,43 @@
 
 v0.3.4 (MM / DD / 2020)
 ------------------------
-Release of new version including time-series albedo options and an updated module.json.** Deprecated support for Python 2.7. Stops support on this release. PVLib >= 0.7.0 is required for metdata file IO **
+Release of new version including time-series albedo options, selecting if weather data is right, left or center labeled (averaged), and an updated module.json.
 
-API Changes
-~~~~~~~~~~~
-* Input changed to :py:class:`~bifacial_radiance.RadianceObj.gendaylit`, now making metdata an optional input.
-* readWeatherData, readEPW and readTMY now have an input ``label``, to mark if data is right-averaged, left-averaged, or center-averaged. Default is ``label=Right``, like TMY3 files: TMY3 data is right-labeled, so 11 AM data represents data from 10 to 11, and sun position is calculated at 10:30 AM. 
+**Deprecated support for Python 2.7. Stops support on this release. PVLib >= 0.7.0 is required for metdata file IO**
 
 Enhancements
 ~~~~~~~~~~~~
 * Enable ``module.json`` to be human readable and sorted. (:issue:`230`)
-HOURLY ALBEDO:  
-* return `MetObj.albedo` from TMY file (:issue:`246`) (:issue:`68`)
-* :py:class:`~bifacial_radiance.RadianceObj.setGround` pulls albedo if readWeatherfile has already been run
-* :py:class:`~bifacial_radiance.RadianceObj.setGround` accepts lists and 3xN RGB arrays
-* Unrealistic albedo values are constrained to [0:1] (:issue:`227`)
+* Albedo:
+    * Albedo: TIME-Varying ALBEDO: now albedo can be given for every timepoint being modeled. 
+    * Albedo: Functions :py:func:`~bifacial_radiance.RadianceObj.readWeatherFile`, :py:func:`~bifacial_radiance.RadianceObj.readEPW`, and :py:func:`~bifacial_radiance.RadianceObj.readTMY` return `MetObj.albedo` from TMY or EPW file if available (:issue:`246`) (:issue:`68`)
+    * Albedo: :py:class:`~bifacial_radiance.RadianceObj.setGround` pulls albedo if readWeatherfile has already been run and no explicit albedo is provided.
+    * Albedo: :py:class:`~bifacial_radiance.RadianceObj.setGround` accepts lists and 3xN RGB arrays
+    * Albedo: Unrealistic albedo values are constrained to [0:1] (:issue:`227`)
+* :py:func:`~bifacial_radiance.RadianceObj.readWeatherFile`, :py:func:`~bifacial_radiance.RadianceObj.readEPW`, and :py:func:`~bifacial_radiance.RadianceObj.readTMY` now have an input ``label``, to mark if data is **right-averaged, left-averaged, or center-averaged**. Default is ``label=Right``, like TMY3 files: TMY3 data is right-labeled, so 11 AM data represents data from 10 to 11, and sun position is calculated at 10:30 AM. Sun position and tracker positions gets calculated according to the label chosen. 
+
+API Changes
+~~~~~~~~~~~
+* Input changed to :py:class:`~bifacial_radiance.RadianceObj.gendaylit`, now making metdata an optional input.
 
 Bug fixes
 ~~~~~~~~~
-* GUI tracker visualization enabled. (:issue:`229`)
+* GUI tracker visualization enabled with proper changes in azimuth. (:issue:`229`)
 * :py:class:`~bifacial_radiance.RadianceObj.makeModule` informational messages updated. (:issue:`202`)
[…]
