### 2020-05-27 8ace27a
upper and lowercase allowed for label.  Add exception for invalid right, left or center input.
