### 2020-05-27 2f575a9
Met obj deltastyle (#253)
* label added to MetObj.  default = right, like TMY
* corrected sunposition for sunrise and susnet for left and right labeled
* Change labelstyle to 'label'; change 'exact' to 'center'; add pytests for right vs left label. Passing.
* comment out quickExample pytest.
* propagating label to readWeatherFile
* whatsnew v 0.3.4 updated (and defaultreadWeatherFile = 'right')
* upper and lowercase allowed for label.  Add exception for invalid right, left or center input.

Co-authored-by: cdeline <chris.deline@nrel.gov>
