### 2020-05-27 2b7030f
Update release date for v0.3.4
