### 2020-05-27 1120422
ipython 7.4 for Sphinx docs
