### 2020-05-28 70f2db1
runtimeWarnings are ignored in set1axis. Fixes pvlib tracker runtimeWarning issue.  Fixes #260
(cherry picked from commit 3b44160f7b6d2f4d89b7c7bf30af9b3facca5b2c)
