### 2020-06-01 621b31b
modscan Analysis1axis option
Modified to address #263 .

If modscan and backscan do not agree on Number of Sensors it will probably crash though. Also modscan has to have the right values/format to work properly.
