### 2020-06-04 f82c006
analysis1axis modscan backscan any number of keys
Adds functionality requested in #263
