### 2020-06-07 #1865 @kyleniemeyer
@shirubana the title and authors for the Zenodo archive should match the paper; actually, it looks like you need to edit the author list as well. (You can edit these things for an existing archive without changing the DOI)
