### 2020-06-07 #1865 @shirubana
Sweet, I think I just did that. I moved the other contributors to contributors and left the author list as in the paper, and changed the title. Hope it's good! Thank you! @kyleniemeyer
