### 2020-06-07 #1865 @shirubana
@kyleniemeyer, there is an image that does not want to render properly on the proof for some reason. I think @melissawm was going to help us investigate, I tried various ways of typing the Markdown to have it show it but could't make it work... any suggestions?
