### 2020-06-07 #1865 @kyleniemeyer
@shirubana I think the issue is that you are using HTML commands for the image, rather than Markdown image commands. Can you replace

```html
<figure>
    <figcaption> Visualization of a bifacial photovoltaic array generated through bifacial_radiance. Courtesy of J. Alderman. </figcaption>
</figure>
```

```markdown
![Visualization of a bifacial photovoltaic array generated through bifacial_radiance. Courtesy of J. Alderman.\label{fig:visualization}](Alderman.PNG)
```

If you want to reference the figure in the text, you can do so with `\autoref{fig: visualization}`
