### 2020-06-07 #1865 @kyleniemeyer
@shirubana just a few final edits to the text:

- in the 4th paragraph ("bifacial_radiance was first coded in ..."), the last sentence uses the references as nouns but uses the usual parenthetical citation. Could you remove the square brackets? For reference, these are the different citation commands and their results:
```
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
```
- please add a period after Stein's middle initial in the `Stein2019b` reference; I think this is why the in-text references are doing "J. S. Stein et al. 2019" and "Stein et al. 2019" (the compiler thinks that they are different people due to the inconsistent period after middle initial)
- in the 5th paragraph ("bifacial_radiance has been used in numerous studies..."), one of the references needs a semicolon rather than a comma: `[@PVSyst, @SAM]`
- You can remove the final sentence that includes a citation to the GitHub repo, since the link to the source code is given on the first page for all JOSS articles
