### 2020-06-07 3f4699e
paper.md updates
in the 4th paragraph ("bifacial_radiance was first coded in ..."), the last sentence uses the references as nouns but uses the usual parenthetical citation. Could you remove the square brackets? For reference, these are the different citation commands and their results:
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
please add a period after Stein's middle initial in the Stein2019b reference; I think this is why the in-text references are doing "J. S
[…]
