+++ paper/paper.bib
@@ -139,7 +139,7 @@
   	year = {2019}}
 
 @inproceedings{Stein2019b,
- 	author = {Stein, Joshua S and Prilliman, Matthew and Stark, Cameron and Nagyvary, John and {Ayala Pelaez}, Silvana and Deline, Chris},
+ 	author = {Stein, Joshua S. and Prilliman, Matthew and Stark, Cameron and Nagyvary, John and {Ayala Pelaez}, Silvana and Deline, Chris},
  	booktitle = {{36th EU PVSEC, Marseille, France}},
  	pages = {1},
  	title = {{Bifacial performance optimization studies using Bifacial Radiance and high performance computing}},
