### 2020-06-08 #1865 @kyleniemeyer
@shirubana almost there! It looks like that reference actually has either a missing or incorrect citation: in `[@PVSyst; @SAM]`, there does not appear to be a corresponding item for `PVSyst` in your bib file.
