+++ paper/paper.bib
@@ -196,4 +196,10 @@
 	url = {http://github.com/NREL/bifacialvf},
 	year = {2019}}
 
-
+@inproceedings{PVSyst,
+ 	author = {Wittmer, Bruno and Mermoud André},
+ 	booktitle = {35th EU PVSEC, Brussels, Belgium},
+ 	title = {{Yield simulations for horizontal axis trackers with bifacial pv modules in PVsyst}},
+ 	url = {https://www.pvsyst.com/wp-content/uploads/2018/10/Poster_PVsyst_2018.pdf},
+ 	year = {2018}}
+ 	
