+++ paper/paper.bib
@@ -170,6 +170,13 @@
 	title = {{View Factor Model and Validation for Bifacial PV and Diffuse Shade on Single-Axis Trackers.}},
 	year = {2017}}
 
+@inproceedings{PVSyst,
+ 	author = {Wittmer, Bruno and Mermoud André},
+ 	booktitle = {35th EU PVSEC, Brussels, Belgium},
+ 	title = {{Yield simulations for horizontal axis trackers with bifacial pv modules in PVsyst}},
+ 	url = {https://www.pvsyst.com/wp-content/uploads/2018/10/Poster_PVsyst_2018.pdf},
+ 	year = {2018}}
+
 @inproceedings{SAM,
 	author = {Blair, Nate and DiOrio, Nicholas and Freeman, Janine and Gilman, Paul and Janzou, Steven and Neises, Ty and Wagner, Michael},
 	institution = {National Renewable Energy Laboratory, Golden, CO},
@@ -196,10 +203,3 @@
 	url = {http://github.com/NREL/bifacialvf},
 	year = {2019}}
 
-@inproceedings{PVSyst,
- 	author = {Wittmer, Bruno and Mermoud André},
- 	booktitle = {35th EU PVSEC, Brussels, Belgium},
- 	title = {{Yield simulations for horizontal axis trackers with bifacial pv modules in PVsyst}},
- 	url = {https://www.pvsyst.com/wp-content/uploads/2018/10/Poster_PVsyst_2018.pdf},
- 	year = {2018}}
- 	
