### 2020-06-08 #1865 @shirubana
@kyleniemeyer  done! I tried to put my lastname in brackets like in the bib { } but the footer didn't recognize that :( So I guess I have to stay with the dash in between. Oh well. Everything else is updated though :)
