### 2020-06-08 #1865 @kyleniemeyer
@openjournals/dev @arfon I thought we could handle double barrel (but no hyphen) last names—is that not the case?
