### 2020-06-08 #1865 @arfon
> @openjournals/dev @arfon I thought we could handle double barrel (but no hyphen) last names—is that not the case?

We can, but we have to handle this manually. Let me know if this is needed here.
