### 2020-06-08 #1865 @arfon
> @arfon that would be very appreciated. Right now the last name reads "Pelaez" instead of "Ayala Pelaez" on the footer :)

Sure thing. I'll fix this once the paper is accepted as it's a manual process.
