### 2020-06-08 #1865 @kyleniemeyer
Hi @shirubana sorry, looks like I missed a few issues in the references:

- remove the "https://doi.org/" from the DOI for the Holmgren 2018 reference
- remove the extraneous `\url{...}` from the Deline 2017 reference, since the DOI link is in another place
- The Anoma 2017 paper is missing the DOI
- The Deline 2019b reference should include the journal and DOI (it looks to be in press now)
- it looks like the SAM reference has a DOI that should be added: 10.2172/1440404
- Also, please look over the references—there are a number that are missing curly brackets around words to enforce capitalization
