+++ paper/paper.bib
@@ -61,14 +61,14 @@
 
 @inproceedings{Ayala2019d,
 	author = {{Ayala Pelaez}, Silvana and Deline, Chris and Stein, Joshua S and Marion, Bill and Anderson, Kevin and Muller, Matthew}, 
-	booktitle = {46th IEEE Photovoltaic Specialists Conference Proceedings, Chicago IL},
+	booktitle = {{46th IEEE Photovoltaic Specialists Conference Proceedings, Chicago IL}},
 	title = {{Effect of torque-tube parameters on rear-irradiance and rear-shading loss for bifacial PV performance on single-axis tracking systems}},
 	url = {https://www.nrel.gov/docs/fy20osti/73203.pdf},
 	year = {2019}}
 
 @inproceedings{Ayala2019e,
  	author = {{Ayala Pelaez}, Silvana and Deline, Chris and MacAlpine, Sara and Olalla, Carlos},
-	booktitle = {6th BifiPV Workshop, Amsterdam, NL},
+	booktitle = {{6th BifiPV Workshop, Amsterdam, NL}},
 	pages = {1},
 	title = {{Bifacial PV system mismatch loss estimation}},
 	url = {https://www.nrel.gov/docs/fy19osti/74831.pdf},
@@ -86,48 +86,53 @@
 
 @inproceedings{Capelle2019,
  	author = {Capelle, Thomas and Araya, Francisco and Haffner, Florent and Sayritupac, Jose and Colin, Herve},
- 	booktitle = {6th BifiPV Workshop, Amsterdam, NL},
+ 	booktitle = {{6th BifiPV Workshop, Amsterdam, NL}},
  	pages = {1--21},
  	title = {{A comparison of bifacial PV system modelling tools}},
  	year = {2019}}
 
 @misc{Deline2017,
 	author = {Deline, Chris and Marion, William and {Ayala Pelaez}, Silvana},
-	howpublished = {[Computer Software] $\backslash$url{\{}https://doi.org/10.11578/dc.20180530.16{\}}},
-	month = {dec},
+	howpublished = {[Computer Software]},
+	month = {Dec},
 	title = {{Bifacial{\_}Radiance}},
-	url = {https://doi.org/10.11578/dc.20180530.16},
+	doi = {10.11578/dc.20180530.16},
 	year = {2017}}
 	
 @inproceedings{Deline2019,
  	author = {Deline, Chris and {Ayala Pelaez}, Silvana and MacAlpine, Sara and Olalla, Carlos},
- 	booktitle = {36th EU PVSEC, Marseille, France},
+ 	booktitle = {{36th EU PVSEC, Marseille, France}},
  	title = {{Bifacial PV system mismatch loss estimation and parameterization}},
  	url = {https://www.nrel.gov/docs/fy20osti/73541.pdf},
+ 	doi = {10.4229/EUPVSEC20192019-5DO.3.4},
+ 	pages = {1449--1453},
+ 	isbn = {3-936338-60-4},
  	year = {2019}}
 
 @article{Deline2019b,
-	author = {Deline, Chris and {Ayala Pelaez}, Silvana and MacAlpine, Sara and Olalla, Carlos},
-	journal = {(Submitted)},
+ 	author = {Deline, Chris and {Ayala Pelaez}, Silvana and MacAlpine, Sara and Olalla, Carlos},
+	doi = {10.1002/pip.3259},
+	journal = {{Progress in Photovoltaics: Research and Applications}},
+	pages = {1--13},
 	title = {{Estimating and parameterizing mismatch power loss in bifacial photovoltaic systems}},
-	year = {2019}}
-
+	year = {2020}}
+	
 @inproceedings{DiOrio2018,
-	author = {Diorio, Nicholas and Deline, Chris},
[…]
