### 2020-06-08 #1865 @kyleniemeyer
Hey @arfon, this is now ready to publish—do you handle the name fix before or after the final step?
