### 2020-06-08 #1865 @shirubana
hold on need to push two more small reference updates! trying to add a DOI for a poster 2 mins
