+++ paper/paper.bib
@@ -78,10 +78,10 @@
 	author = {Berrian, Djaber and Libal, Joris and Klenk, Markus and Nussbaumer, Hartmut and Kopecek, Radovan},
 	doi = {10.1109/JPHOTOV.2019.2924394},
 	journal = {IEEE Journal of Photovoltaics},
-	pages = {2156--3381},
+	pages = {1583--1589},
 	publisher = {IEEE},
 	title = {{Performance of bifacial PV arrays with fixed tilt and horizontal single-axis tracking: Comparison of simulated and measured data}},
-	volume = {Early Acce},
+	volume = {9},
 	year = {2019}}
 
 @inproceedings{Capelle2019,
@@ -93,9 +93,9 @@
 
 @misc{Deline2017,
 	author = {Deline, Chris and Marion, William and {Ayala Pelaez}, Silvana},
-	howpublished = {[Computer Software]},
+	howpublished = {[Computer software]},
 	month = {Dec},
-	title = {{Bifacial{\_}Radiance}},
+	title = {{bifacial{\_}radiance}},
 	doi = {10.11578/dc.20180530.16},
 	year = {2017}}
 	
@@ -147,7 +147,8 @@
  	author = {Stein, Joshua S. and Prilliman, Matthew and Stark, Cameron and Nagyvary, John and {Ayala Pelaez}, Silvana and Deline, Chris},
  	booktitle = {{36th EU PVSEC, Marseille, France}},
  	pages = {1},
- 	title = {{Bifacial performance optimization studies using Bifacial Radiance and high performance computing}},
+ 	title = {{Bifacial performance optimization studies using bifacial_radiance and high performance computing}},
+ 	isbn = {3-936338-60-4},
  	year = {2019}}
 
 @inproceedings{Ward1994,
@@ -171,7 +172,7 @@
 	
 @inproceedings{PVfactors,
 	author = {Anoma, Marc A. and Jacob, David and Bourne, Ben C. and Scholl, Jonathan A., and Riley, Daniel M. and Hansen, Clifford W.}, 
-	booktitle = {44th IEEE Photovoltaic Specialists Conference Proceedings, Washington, DC},
+	booktitle = {{44th IEEE Photovoltaic Specialists Conference Proceedings, Washington, DC}},
 	title = {{View factor model and validation for bifacial PV and diffuse shade on single-axis trackers}},
 	doi = {10.1109/PVSC.2017.8366704},
 	year = {2017}}
@@ -189,7 +190,7 @@
 	pages = {1--19},
 	publisher = {NREL/TP-6A20-70414},
 	doi = {10.2172/1440404},
-	title = {{System Advisor Model (SAM) General Description (Version 2017.9.5}},
+	title = {{System Advisor Model (SAM) general description (version 2017.9.5)}},
 	year = {2018}}
 
 @inproceedings{Marion2017,
