+++ paper/paper.bib
@@ -147,7 +147,7 @@
  	author = {Stein, Joshua S. and Prilliman, Matthew and Stark, Cameron and Nagyvary, John and {Ayala Pelaez}, Silvana and Deline, Chris},
  	booktitle = {{36th EU PVSEC, Marseille, France}},
  	pages = {1},
- 	title = {{Bifacial performance optimization studies using bifacial_radiance and high performance computing}},
+ 	title = {{Bifacial performance optimization studies using bifacial{\_}radiance and high performance computing}},
  	isbn = {3-936338-60-4},
  	year = {2019}}
 
