+++ paper/paper.bib
@@ -182,6 +182,7 @@
  	booktitle = {{35th EU PVSEC, Brussels, Belgium}},
  	title = {{Yield simulations for horizontal axis trackers with bifacial pv modules in PVsyst}},
  	url = {https://www.pvsyst.com/wp-content/uploads/2018/10/Poster_PVsyst_2018.pdf},
+ 	doi = {10.4229/35thEUPVSEC20182018-6CV.2.19},
  	year = {2018}}
 
 @inproceedings{SAM,
