### 2020-06-08 #1865 @shirubana
@kyleniemeyer ok I'm good , all references have DOI if they have them now.
