### 2020-06-08 #1865 @kyleniemeyer
Hey @arfon, this is now ready for the name fix, both in the paper footer and the citation on the landing page.
